### 2026-01-14 #9763 @Digdgeo
Hi again,

After seeing the reviewers list, I'd like to add two more suggestions:
- **Aaron Zuspan** (aazuspan) - wxee author
- **Kel Markert** (kmarkert) - GEE remote sensing expert
