### 2026-01-21 2beae42
Update paper for JOSS new requirements: add Software Design, Research Impact, and AI Disclosure sections
