+++ paper/paper.bib
@@ -70,3 +70,22 @@
   note={Technical Report, eLTER Plus Project},
   url={https://www.researchgate.net/publication/396609384}
 }
+
+@article{Garcia2023,
+  title={PhenoApp: A Google Earth Engine based tool for monitoring phenology},
+  author={García-Díaz, Diego and Díaz-Delgado, Ricardo},
+  journal={Revista de Teledetección},
+  number={61},
+  pages={73--81},
+  year={2023},
+  doi={10.4995/raet.2023.18767}
+}
+
+@techreport{DiazDelgado2024b,
+  title={eLTER PLUS Deliverable D4.4 - Piloting of eLTER specific downstream services contributing to eLTER Standard Observation Variables},
+  author={Díaz-Delgado, Ricardo and García-Díaz, Diego and Marangi, Catia and Silver, Michael and Oggioni, Alessandro and Tagliolato Acquaviva D'Aragona, Paolo and Karnieli, Arnon and Peterseil, Johannes and Vicario, Saverio and Anttila, Sanna and Böttcher, Kristin and Bolton, William and Wohner, Christoph and Minic, Vojislava},
+  institution={eLTER PLUS Project},
+  year={2024},
+  note={EC Accepted Version},
+  doi={10.5281/zenodo.14889212}
+}
