+++ paper/paper.md
@@ -87,11 +87,7 @@ ndvi = NdviSeasonality(roi=roi, start_year=2013, end_year=2013,
 collection = ndvi.get_year_composite()  # ImageCollection: 1 image × 4 bands
 ```
 
-This simplification extends across 7 satellite platforms, 88 spectral and climate variables, multiple temporal aggregation methods, and flexible region of interest specifications—all through a unified API that inherits geemap's interactive mapping capabilities.
-
-Several Python packages extend GEE for remote sensing workflows with similar approaches to code simplification: eemont [@Montero2021] provides preprocessing and spectral indices computation, while wxee [@Zuspan2021] enables temporal aggregation and xarray export. These tools can be combined for similar tasks. However, Ndvi2Gif (first released in May 2020) differentiates itself by organizing temporal data within `ee.ImageCollection` objects differently. While eemont maintains the original temporal structure and wxee aggregates into sequential time steps (e.g., 480 monthly composites for 40 years), ndvi2gif organizes data hierarchically where **each image represents a year with bands per period** (e.g., 40 images/years with 12, 24, or 52 bands each for monthly, biweekly, or weekly composites). This band-based organization enables direct multi-decadal queries such as: *What is the mean cyanobacteria index (NDCI) detected from Sentinel-2 in Andalusian reservoirs during August across the last 10 years?* or *In which week of the year does chlorophyll (OCI) detected from Sentinel-3 peak in European lakes and what is the maximum value reached?*—rather than requiring temporal filtering, grouping, and reduction operations across hundreds of separate images.
-
-Beyond this organizational approach, ndvi2gif provides integrated capabilities unavailable in other GEE Python packages: complete Sentinel-1 Analysis Ready Data (ARD) preprocessing based on Vollrath et al. [-@Vollrath2020], automated phenology metrics extraction (Start/End/Peak of Season with trend analysis), machine learning classification, zonal statistics via `get_stats()`, and geemap-based visualization—all within a unified API built around the band-organized composite structure. The package has been applied in calibration and validation studies of Earth Observation products, including phenology and surface water monitoring at the Doñana protected area [@DiazDelgado2024].
+This simplification extends across 7 satellite platforms, 88 spectral and climate variables, multiple temporal aggregation methods, and flexible region of interest specifications—all through a unified API that inherits geemap's interactive mapping capabilities. The architectural decisions and design philosophy that enable this simplification, along with how Ndvi2Gif differentiates from other GEE Python packages, are detailed in the Software Design section below.
 
 # Key features
 
@@ -150,6 +146,26 @@ The `TimeSeriesAnalyzer` module extracts temporal patterns from composite stacks
 ## Machine learning classification
 
 The `LandCoverClassifier` module provides supervised (Random Forest, SVM, CART, Gradient Tree Boost) and unsupervised (K-means, Cascade K-means) classification with accuracy assessment and feature importance analysis.
+
+# Software Design
+
+Several Python packages extend GEE for remote sensing workflows: eemont [@Montero2021] provides preprocessing and spectral indices computation, while wxee [@Zuspan2021] enables temporal aggregation and xarray export. These tools can be combined for similar tasks. However, Ndvi2Gif (first released in May 2020) adopts a fundamentally different temporal organization strategy. While eemont maintains the native temporal structure of `ee.ImageCollection` objects and wxee aggregates time series into sequential steps (e.g., 480 monthly composites for 40 years), ndvi2gif implements a hierarchical band-based approach where **each image represents a temporal unit (typically a year) with bands encoding sub-periods** (e.g., 40 images/years with 12, 24, or 52 bands each for monthly, biweekly, or weekly composites).
+
[…]
