### 2026-01-21 #9763 @Digdgeo
Hi @kthyng,

I have now updated the paper to comply with the new format and scope criteria.

## Changes made:

### 1. **Software Design section** (~250 words)
- Detailed comparison with existing GEE Python packages (eemont, wxee)
- Explanation of the band-based temporal organization architecture
- Concrete query examples demonstrating the design advantages
- "Build vs. contribute" justification

### 2. **Research Impact Statement section** (~200 words)
- Documented realized impact:
  - Peer-reviewed publication on Doñana validation studies
  - Foundation for PhenoApp (published in Revista de Teledetección)
  - Core component of geeltermap (officially documented in eLTER PLUS Deliverable D4.4)
  - Integration in eLTER and SUMHAL operational projects
- Technical significance demonstrating research-readiness

### 3. **AI Usage Disclosure section**
- Full transparency on AI tool usage in development workflow
- Clear distinction between AI assistance and intellectual contributions

The updated paper.md and paper.bib files are now available in the repository.

**Additional improvement:** I am currently working on a visual diagram illustrating the library's architecture and temporal organization approach, which I plan to add to the paper in the coming days to enhance clarity.

Please let me know if any further modifications are needed.
