### 2026-01-22 a11c2b4
Add workflow diagram and updated paper for JOSS review
