### 2026-01-22 #9763 @Digdgeo
Hi @kthyng,

I have now completed the visual diagram I mentioned in my previous comment.

The workflow architecture figure has been added to the paper (Figure 1) and is now available in the repository at `paper/ndvi2gif_workflow.png`. This diagram illustrates:
- The multi-platform data input system
- The band-based temporal organization architecture
- The main analysis classes and their relationships
- The integrated workflow from data ingestion to final outputs

I will continue updating the tutorial section and adding extra functionalities to the library in the near future, but I think that the paper is ready for publication. I look forward to the reviewers' feedback.

Thanks a lot,
Diego
