### 2026-01-28 de5e113
Update JOSS paper: add State of the Field section and fix references
- Add State of the Field section comparing with existing GEE tools
- Fix figure label reference (fig → fig:workflow)
- Fix table reference format (Table \autoref → \autoref)
- Update Summary with accurate counts: 7 satellite platforms, 52 spectral/radar indices, 48 climate variables
- Simplify figure caption with color coding legend
- Expand acknowledgements for geemap contribution
