+++ paper/paper.md
@@ -1,28 +1,31 @@
 ---
+
 title: 'Ndvi2Gif: A Python Package for Multi-Seasonal Remote Sensing Analysis via Google Earth Engine'
 tags:
-  - Python
-  - remote sensing
-  - Google Earth Engine
-  - time series analysis
-  - vegetation monitoring
-  - SAR processing
-  - phenology
-  - NDVI
-authors:
-  - name: Diego García Díaz
-    orcid: 0000-0002-2757-7112
-    affiliation: 1
-affiliations:
- - name: Estación Biológica de Doñana (EBD-CSIC), Spain
-   index: 1
-date: 28 December 2025
-bibliography: paper.bib
+
+* Python
+* remote sensing
+* Google Earth Engine
+* time series analysis
+* vegetation monitoring
+* SAR processing
+* phenology
+* NDVI
+  authors:
+* name: Diego García Díaz
+  orcid: 0000-0002-2757-7112
+  affiliation: 1
+  affiliations:
+* name: Estación Biológica de Doñana (EBD-CSIC), Spain
+  index: 1
+  date: 28 December 2025
+  bibliography: paper.bib
+
 ---
 
 # Summary
 
-Google Earth Engine (GEE) is a cloud-based platform for planetary-scale geospatial analysis [@Gorelick2017]. Ndvi2Gif is a Python package built on top of GEE and geemap [@Wu2020] that simplifies the generation and analysis of multi-temporal composite images from satellite and climate reanalysis data.
+Google Earth Engine (GEE) is a cloud-based platform for planetary-scale geospatial analysis [@Gorelick2017]. Ndvi2Gif is a Python package built on top of GEE and geemap [@Wu2020] that simplifies the generation and analysis of multi-temporal composite images from satellite and climate reanalysis data. The package provides unified access to 7 satellite platforms (Sentinel-1/2/3, Landsat 4–9, MODIS) and climate reanalysis data (ERA5-Land, CHIRPS), with 52 predefined spectral and radar indices plus 48 climate variables.
 
 The core concept of Ndvi2Gif is the creation of **multi-seasonal statistical composites**. The `get_year_composite()` method returns an `ee.ImageCollection` where each image represents one year, with bands for each temporal period (e.g., `periods=12` generates bands named `january`, `february`, ..., `december`). This band-based temporal organization enables queries such as: *What was the maximum flood extent detected in January across 40 years of Landsat (1984–2024)?* by simply calling `collection.select('january').max()`. Built-in zonal statistics extraction via `get_stats()` allows direct computation of statistics per polygon without external GIS processing.
 
@@ -82,14 +85,24 @@ The 40 lines above can be reduced to 5 lines using Ndvi2Gif:
 ```python
 from ndvi2gif import NdviSeasonality
 
-ndvi = NdviSeasonality(roi=roi, start_year=2013, end_year=2013, 
+ndvi = NdviSeasonality(roi=roi, start_year=2013, end_year=2013,
                        sat='Landsat', index='ndvi', periods=4)
-collection = ndvi.get_year_composite()  # ImageCollection: 1 image × 4 bands
+collection = ndvi.get_year_composite()
 ```
 
-This simplification extends across 7 satellite platforms, 88 spectral and climate variables, multiple temporal aggregation methods, and flexible region of interest specifications—all through a unified API that inherits geemap's interactive mapping capabilities. The architectural decisions and design philosophy that enable this simplification, along with how Ndvi2Gif differentiates from other GEE Python packages, are detailed in the Software Design section below. \autoref{fig:workflow} provides an overview of the package's workflow and capabilities.
[…]
