### 2026-01-28 cfc9e78
Fix YAML frontmatter formatting
