### 2026-01-28 #9763 @Digdgeo
Hi @kthyng @arfon,

Sorry, my fault. I should have noticed that the bot already mentioned the missing "State of the Field" section in its checks.

I have now added this missing section.

The paper now includes all four required sections:
- State of the Field
- Software Design
- Research Impact Statement
- AI Usage Disclosure

The updated paper.md and paper.bib are available in the repository. Please let me know if any further modifications are needed.

Thanks again :)
