### 2026-02-16 #9763 @Digdgeo
Hi @kthyng, just a friendly follow-up. It's been a couple of weeks since the last repository check and all sections are now passing. I wanted to check if there's any update on the scope evaluation or if there's anything else you need from my side.
Also, I noticed the wordcount warning (1917 words). If that's a concern, I could easily trim it down — for instance, the code snippet in the paper is quite similar to the one in the eemont paper, so I could shorten that section without losing clarity.
Thanks for your time!
