### 2026-03-05 8bd90d3
Set system:time_start on yearly composites in get_year_composite
ee.Image.cat() does not propagate system:time_start, so images in
the resulting collection had no date metadata. Now each composite
is stamped with the corresponding year so that img.date(),
collection.filterDate() and similar EE operations work correctly.
