### 2026-03-06 0c914e1
Add FAI index and bump version to 1.1.0
Add Floating Algae Index (Hu 2009) for cyanobacterial bloom detection
in water bodies, with sensor-specific NIR baseline wavelength factors
for Sentinel-2, Landsat (4-9), and MODIS. Update README and version.
