+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ndvi2gif
-version = 1.0.0
+version = 1.1.0
 description = Multi-seasonal remote sensing analysis suite: composites, SAR preprocessing, time series analytics, and land cover classification with Google Earth Engine
 long_description = file: README.md
 long_description_content_type = text/markdown
