### 2026-03-09 #9763 @Digdgeo
Hi @kthyng and @arfon, apologies — I don't want to be a bother, but I'm just a bit surprised and wasn't sure if it's normal to have gone several weeks without any news on this. I also wanted to mention that a couple of days ago I updated the package to version 1.1.0 (I simply added a new index and a new notebook following a query I had at work), and I wasn't sure if that requires any changes on the submission side. I also wanted to check back on the bot review from February 3rd, where everything passed correctly and there was only a warning about the word count (1917). I wasn't sure if I should address that or if it's fine as it is. Thank you so much for your time!
