### 2026-03-17 #9763 @kthyng
@Digdgeo Sorry for the delay. All the changes on our end lead to every task taking extra time for now as we learn the new procedures, but without me necessarily having extra time to spend on it, so I am pretty behind.

That said, here are the high level checks I need to go through to process your submission at a high level:

Hi and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment over time as I have a chance to go through them:

- [x]  [Not a minor utility](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance)

New submission criteria since journal scope changes in January 2026 (more details in [submission docs](https://joss.readthedocs.io/en/latest/submitting.html#pre-review-screening-criteria)):
- [x] Public development history ≥ 6 months
- [x] Demonstrated research impact
- [x] Open source practices
- [x] Iterative development
- [x] Check for borderline cases
