### 2026-03-17 #9763 @kthyng
@Digdgeo
* Yes would be better to shorten your paper a bit
* I can update the version of your submission
