### 2026-03-17 #9763 @Digdgeo
Hi @kthyng,

Thank you so much for the update and for taking the time to go through the checklist — I really appreciate it, and I fully understand that the recent changes have added extra work on your end.

I will work on shortening the paper this weekend to bring it within the 750–1750 word limit. I will also put together a longer list of suggested reviewers (up to 5) and share it here once ready.

Thanks again for your patience and support!

Best regards,
Diego
