### 2026-03-17 #9763 @kthyng
@Digdgeo

I would recommend you shorten your README - it was a lot to get through! I would put the most important items in there and the rest in the docs. This is just a suggestion.
