### 2026-03-22 ebc349b
Trim paper to JOSS word limit and update README
- Remove 40-line boilerplate code block from Statement of Need
- Condense Software Design section (removed repetition with State of the Field)
- Condense Time Series and ML sections
- Add missing DOI (10.62880/rars25009) to DiazDelgado2024 bib entry, convert to @techreport
- README: remove outdated v0.6.0 release notes, roadmap, duplicate docs section, condense Quick Usage Example
