+++ paper/paper.bib
@@ -63,12 +63,12 @@
   year={2021}
 }
 
-@misc{DiazDelgado2024,
+@techreport{DiazDelgado2024,
   title={In-situ validation of Land Surface Phenology, Land Surface Temperature and Surface Water derived from Earth Observation products: Doñana protected area as a potential cal/val supersite},
   author={Díaz-Delgado, Ricardo and García-Díaz, Diego},
+  institution={eLTER Plus Project},
   year={2024},
-  note={Technical Report, eLTER Plus Project},
-  url={https://www.researchgate.net/publication/396609384}
+  doi={10.62880/rars25009}
 }
 
 @article{Garcia2023,
