+++ paper/paper.md
@@ -28,56 +28,7 @@ The core concept of Ndvi2Gif is the creation of **multi-seasonal statistical com
 
 # Statement of need
 
-Generating multi-temporal composites in GEE requires extensive code for collection filtering, cloud masking, scaling to reflectance, and temporal aggregation. The following example shows the standard approach to create seasonal NDVI composites from Landsat:
-
-```python
-import ee
-ee.Initialize()
-
-roi = ee.Geometry.Point([-5.9, 37.2]).buffer(10000)
-
-# Scale Landsat 8-9 (OLI) to surface reflectance
-def scale_OLI(image):
-    optical = image.select(['SR_B2','SR_B3','SR_B4','SR_B5','SR_B6','SR_B7']) \
-                   .multiply(0.0000275).add(-0.2) \
-                   .rename(['Blue','Green','Red','Nir','Swir1','Swir2'])
-    return image.addBands(optical, None, True)
-
-# Scale Landsat 4-7 (TM/ETM+) to surface reflectance  
-def scale_ETM(image):
-    optical = image.select(['SR_B1','SR_B2','SR_B3','SR_B4','SR_B5','SR_B7']) \
-                   .multiply(0.0000275).add(-0.2) \
-                   .rename(['Blue','Green','Red','Nir','Swir1','Swir2'])
-    return image.addBands(optical, None, True)
-
-# Cloud mask using QA_PIXEL band
-def maskClouds(image):
-    qa = image.select('QA_PIXEL')
-    mask = qa.bitwiseAnd(1 << 3).eq(0).And(qa.bitwiseAnd(1 << 4).eq(0))
-    return image.updateMask(mask)
-
-def addNDVI(image):
-    return image.addBands(
-        image.normalizedDifference(['Nir', 'Red']).rename('NDVI'))
-
-def getSeasonalComposite(startDate, endDate):
-    l8 = ee.ImageCollection('LANDSAT/LC08/C02/T1_L2') \
-           .filterBounds(roi).filterDate(startDate, endDate) \
-           .map(maskClouds).map(scale_OLI).map(addNDVI)
-    l5 = ee.ImageCollection('LANDSAT/LT05/C02/T1_L2') \
-           .filterBounds(roi).filterDate(startDate, endDate) \
-           .map(maskClouds).map(scale_ETM).map(addNDVI)
-    return l8.merge(l5).select('NDVI').max()
-
-composite = ee.Image.cat([
-    getSeasonalComposite('2013-01-01', '2013-03-31'),
-    getSeasonalComposite('2013-04-01', '2013-06-30'),
-    getSeasonalComposite('2013-07-01', '2013-09-30'),
-    getSeasonalComposite('2013-10-01', '2013-12-31')
-])
-```
-
-The 40 lines above can be reduced to 5 lines using Ndvi2Gif:
+Generating multi-temporal composites in GEE requires 30–40 lines of boilerplate per sensor for collection filtering, cloud masking, scaling to reflectance, and temporal aggregation. Using Ndvi2Gif, this reduces to:
 
 ```python
 from ndvi2gif import NdviSeasonality
@@ -149,7 +100,7 @@ This preprocessing enables the reliable use of SAR-derived indices for vegetatio
 
 ## Time series and phenology analysis
 
-The `TimeSeriesAnalyzer` module provides a comprehensive framework for extracting, analysing, and visualising temporal dynamics from the multi-seasonal composite stacks generated by `NdviSeasonality`. Unlike pixel-level time series derived directly from raw `ee.ImageCollection` objects that encode time implicitly as a sequence of images, this approach operates on temporally aligned composite periods stored explicitly as bands within each yearly image, ensuring consistency across years and reducing sensitivity to irregular acquisition frequency or missing observations.
[…]
