+++ README.md
@@ -43,65 +43,13 @@ Ndvi2Gif was updated and extended as part of its integration into the eLTER and 
 
 ---
 
-## ✨ What's New in v1.1.0
+## ✨ What's New in v1.0.0 / v1.1.0
 
-The **1.1.0 release** adds the **Floating Algae Index (FAI)** for multi-sensor cyanobacterial bloom and floating algae detection, with sensor-specific band wavelength calibration.
-
-### 🌊 Key Features in v1.1.0
-
-- **FAI (Floating Algae Index)**: Original Hu (2009) formulation using surface reflectance. Detects floating algae and cyanobacterial bloom accumulations in water bodies. Available for Sentinel-2, Landsat (4–9), and MODIS with sensor-specific NIR baseline wavelength factors.
-
----
-
-## ✨ What's New in v1.0.0 - First Stable Release 🎉
-
-The **1.0.0 release** marks *Ndvi2Gif* as production-ready, expanding beyond vegetation monitoring into comprehensive **climate analysis** with ERA5-Land reanalysis (47 variables, 1950-present) and CHIRPS precipitation (1981-present). The library now supports **88 variables across 7 platforms** with intelligent handling of climate vs. vegetation data in time series analysis.
-
-**NEW in v1.0.0:** Complete Jupyter Book documentation for JOSS (Journal of Open Source Software) submission.
-
-### 🌡️ Key Features in v1.0.0
-
-- **ERA5-Land Climate Reanalysis**: 47 variables including temperature (with min/max/Celsius variants), precipitation (meters and L/m²), soil moisture, radiation, wind, and snow
-- **CHIRPS Precipitation**: High-resolution daily rainfall (1981-present, ~5.5km) combining satellite + station data
-- **Enhanced Statistics**: New `sum` and `min` aggregators for climate data
-- **Smart Time Series**: Climate-specific dashboards with seasonal statistics (replaces phenology for non-vegetation data)
-- **Bug Fixes**: `end_year` now inclusive, Sentinel-3 band naming fixed, ROI centroid error resolved
-
----
-
-## 🧠 Previous Release: v0.6.0 - Machine Learning & Classification
-
-The **0.6.0 release** transformed *Ndvi2Gif* into a **complete remote sensing analysis platform** with integrated machine learning capabilities.
-
-### 🚀 New Classification Capabilities
-
-- 🧠 **LandCoverClassifier** – Complete supervised and unsupervised classification workflows
-- 🎯 **Multiple Algorithms** – Random Forest, SVM, CART, Naive Bayes, Gradient Tree Boost, K-means, LDA
-- 📊 **Accuracy Assessment** – Confusion matrices and comprehensive validation reports
-- 🔧 **Feature Engineering** – Multi-temporal stacks with automatic normalization
-- 📤 **Enhanced Exports** – Direct export to Google Drive and Earth Engine Assets
-
-### 📚 Documentation Overhaul
-
-- **95%+ Documentation Coverage** – Comprehensive Sphinx-style docstrings
-- **Scientific References** – All indices now include citations
-- **Complete Examples** – Every method includes usage examples
-- **Better Error Handling** – Informative messages with suggested solutions
-
-### 🛰️ Enhanced Capabilities
-
-- **export_to_drive()** – Batch export with full parameter control
-- **export_to_asset()** – Direct Earth Engine Asset creation with pyramiding policies
-- **Automatic scale detection** – Sensor-specific resolution handling
[…]
