### 2026-03-22 #9763 @Digdgeo
Hi @kthyng, thank you for your feedback! (and sorry for the delay, busy week here)

I have addressed the main pending items:

  - Word count: The paper has been trimmed to comply with the 750–1750 word limit. The long boilerplate code block in the Statement of Need has been removed, and the Software Design section has been condensed to avoid repetition with the State of the Field section.
  - Missing DOI: Added 10.62880/rars25009 to the Díaz-Delgado & García-Díaz (2024) reference.
  - README: Shortened as suggested — removed outdated release notes, duplicate documentation section, and condensed the usage examples.

Regarding suggested reviewers, here are the five candidates I propose:

  1. davemlz — David Montero-Loaiza, author of eemont
  2. aazuspan — Aaron Zuspan, author of wxee
  3. kmarkert — Kel Markert, GEE remote sensing expert
  4. samvanstroud — author of pyveg
  5. YOUNKAP — remote sensing and GEE expertise

Finally, I wanted to mention that I am currently preparing two additional example notebooks for the documentation that I expect to publish this week: one on water quality monitoring in reservoirs using Sentinel-2 and Sentinel-3, and another on vegetation classification in highly cloudy regions combining Sentinel-1 SAR and Sentinel-2 imagery.
