### 2026-03-23 #9763 @kthyng
@Digdgeo You might still shorten that readme. Often they have high level highlights of capabilities, installation instructions, citation, describing how someone can get involved, maybe a little bit more, plus a link to docs where all the rest of the info is. It is really quite a long readme.
