+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ndvi2gif
-version = 1.1.0
+version = 1.2.0
 description = Multi-seasonal remote sensing analysis suite: composites, SAR preprocessing, time series analytics, and land cover classification with Google Earth Engine
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -38,7 +38,7 @@ python_requires = >=3.8
 install_requires =
     geemap>=0.29.5
     earthengine-api>=0.1.347
-    numpy<2.0
+    numpy>=1.24
     pandas
     scipy
     matplotlib
