### 2026-04-12 f842913
Bump version to 1.2.0: add HydroperiodAnalyzer, SCL masking, fix numpy pin
- Add HydroperiodAnalyzer module (GEE-native hydroperiod analysis)
- Add scl_mask parameter to NdviSeasonality (SCL-based cloud masking for S2)
- Fix numpy dependency: numpy<2.0 → numpy>=1.24 (numpy 2.x supported)
