+++ ndvi2gif/__init__.py
@@ -7,12 +7,13 @@ with Google Earth Engine.
 
 __author__ = "Diego García Díaz"
 __email__ = "diegogarcia@ebd.csic.es"
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 from .ndvi2gif import NdviSeasonality, scale_OLI, scale_ETM
 from .s1_ard import S1ARDProcessor
 from .timeseries import TimeSeriesAnalyzer, SpatialTrendAnalyzer
 from .clasification import LandCoverClassifier
+from .hydroperiod import HydroperiodAnalyzer
 
 __all__ = [
     "NdviSeasonality",
@@ -20,6 +21,7 @@ __all__ = [
     "TimeSeriesAnalyzer",
     "SpatialTrendAnalyzer",
     "LandCoverClassifier",
+    "HydroperiodAnalyzer",
     "scale_OLI",
     "scale_ETM",
 ]
