### 2026-04-12 ab08ea0
Update README and CHANGELOG for v1.2.0
Add HydroperiodAnalyzer section to README (with phydroperiod credit),
update What's New section, and add v1.2.0 CHANGELOG entry.
