### 2026-04-12 fa0793b
Trim README: concise structure, fix citation year, acknowledgments
