+++ README.md
@@ -9,504 +9,124 @@
 ![NDVI2GIF Köln](https://i.imgur.com/Y5dOWIk.jpeg)
 *Richter's stained glass in Cologne Cathedral. Inspiration for this library.*
 
-**Ndvi2Gif** is a Python library designed to simplify access to global satellite data through the Google Earth Engine platform. While its name highlights the ability to create seasonal GIF animations, the true power of this tool lies in its capability to compute and export pixel-wise statistics for any region on Earth, across any time span covered by supported remote sensing datasets.
+**Ndvi2Gif** is a Python library for multi-temporal remote sensing analysis with Google Earth Engine. It provides seasonal compositing, 40+ vegetation and environmental indices, SAR preprocessing, time series analytics, land cover classification, and hydroperiod analysis — all server-side, without downloading raw data.
 
-Built on top of [Google Earth Engine](https://github.com/google/earthengine-api) and [Geemap](https://github.com/giswqs/geemap), it allows you to:
+Built on top of [Google Earth Engine](https://github.com/google/earthengine-api) and [geemap](https://github.com/giswqs/geemap). Adapted and extended through its use in the [eLTER](https://elter-ri.eu/) and [SUMHAL](https://lifewatcheric-sumhal.csic.es/) projects.
 
-- Generate annual or multi-annual composited rasters (e.g., median NDVI per season between 2001 and 2020),
-- Apply multiple statistics (mean, max, flexible percentiles) across space and time,
-- Export results as GeoTIFFs for further analysis,
-- Retrieve zonal statistics over user-defined geometries,
-- Monitor vegetation structure with advanced SAR indices,
-- Handle incomplete years automatically for real-time monitoring,
-- **NEW in v0.6.0:** Perform supervised and unsupervised land cover classification with integrated machine learning,
-- **NEW in v0.6.0:** Export directly to Google Drive and Earth Engine Assets,
-- And yes — also create colorful GIFs for easy visualization.
-
-Whether you're monitoring crop phenology, detecting harvest events, assessing drought trends, classifying land cover, or preparing input layers for further ecological modeling, `ndvi2gif` makes it easier to extract reliable, multi-temporal remote sensing information at scale.
-
-Ndvi2Gif was updated and extended as part of its integration into the eLTER and SUMHAL projects, which also enabled the use of eLTER site boundaries (via `deimsPy`) as one of its input sources.
-
-![Interface Screenshot](https://i.imgur.com/Sv9LfYj.png)
+---
 
 ## 📚 Documentation
 
-**Complete documentation is now available as an interactive Jupyter Book:**
+**https://digdgeo.github.io/Ndvi2Gif/**
 
-🔗 **https://digdgeo.github.io/Ndvi2Gif/**
-
-- [Getting Started Guide](https://digdgeo.github.io/Ndvi2Gif/getting_started/installation.html) - Installation and authentication
-- [API Reference](https://digdgeo.github.io/Ndvi2Gif/reference/api.html) - Complete class and method documentation
-- [Datasets Guide](https://digdgeo.github.io/Ndvi2Gif/reference/datasets.html) - Detailed specs for all 7 platforms
-- [Indices Catalog](https://digdgeo.github.io/Ndvi2Gif/reference/indices.html) - Reference for all 88 variables
[…]
