### 2026-04-12 eb5bac2
Update conda recipe to v1.2.0: fix numpy pin, add missing deps, update description
