### 2026-04-12 #9763 @Digdgeo
Hi there @mengqi-z, just a quick update: I've significantly shortened the README as suggested previously, and I've also released a new version v1.2.2 (sorry, a couple of silly bugs (as usual) in versions 1.2.0 and 1.2.1) which adds a new module for Hydroperiod calculation based on the flood seasonal composites. Happy to update the version tag if needed. Looking forward to hearing from you!
