### 2026-04-12 9fd82ee
Fix packaging: remove py_modules, exclude tests, bump to 1.2.1
