+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ndvi2gif
-version = 1.2.0
+version = 1.2.1
 description = Multi-seasonal remote sensing analysis suite: composites, SAR preprocessing, time series analytics, and land cover classification with Google Earth Engine
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -28,11 +28,6 @@ classifiers =
 
 [options]
 packages = find:
-py_modules =
-    ndvi2gif
-    s1_ard
-    timeseries
-    clasification
 
 python_requires = >=3.8
 install_requires =
@@ -55,6 +50,10 @@ install_requires =
     scikit-learn>=1.0
 
 include_package_data = True
+
+[options.packages.find]
+exclude =
+    tests*
 
 [options.extras_require]
 dev =
