### 2026-04-12 989b25b
Add hydroperiod.py to repo and bump to 1.2.2
