+++ ndvi2gif/hydroperiod.py
@@ -1,0 +1,927 @@
+"""
+hydroperiod.py - GEE-native Hydroperiod Analysis
+=================================================
+
+Computes hydroperiod (flood days per hydrological year) entirely in Google
+Earth Engine using the midpoint temporal weighting method.
+
+The hydroperiod represents the number of days a pixel remains flooded during
+a hydrological cycle (default: September 1st – August 31st).
+
+Methodology
+-----------
+The midpoint weighting method distributes 365 days among available scenes:
+each scene is assigned a temporal "territory" that extends from the midpoint
+with its predecessor to the midpoint with its successor (boundaries at day 0
+and day 365 for the first and last scenes). Weighted flood and valid days are
+then accumulated per pixel and optionally normalised to a full year.
+
+References
+----------
+Bustamante, J., Aragonés, D., Afán, I., Luque, C.J., Pérez-Vázquez, A.,
+Castellanos, E.M., Díaz-Delgado, R. (2016). Predictive models of floristic
+composition based on flooding frequency in Mediterranean wetlands.
+Remote Sensing, 8(9), 776.
+
+Author: Diego García Díaz
+"""
+
+import ee
+import numpy as np
+from datetime import date
+
+
+class HydroperiodAnalyzer:
+    """GEE-native hydroperiod computation from satellite water indices.
+
+    Computes hydroperiod entirely in Google Earth Engine using the midpoint
+    temporal weighting method. Each scene in the satellite collection receives
+    a proportional number of days based on its position within the hydrological
+    cycle, then weighted flood and valid days are accumulated per pixel.
+
+    Parameters
+    ----------
+    ndvi_seasonality : NdviSeasonality
+        Configured :class:`NdviSeasonality` instance. The ROI, satellite,
+        date range, cloud filtering, and band standardisation are reused
+        directly — no duplicate configuration needed.
+    hydrological_year_start : tuple of (int, int), optional
+        ``(month, day)`` marking the start of the hydrological cycle.
+        Default ``(9, 1)`` = September 1st.
+
+    Examples
+    --------
+    ::
+
+        import ee
+        from ndvi2gif import NdviSeasonality, HydroperiodAnalyzer
+
+        ee.Initialize()
+
+        ns = NdviSeasonality(roi, sat='S2', start_year=2023, end_year=2024)
+        ha = HydroperiodAnalyzer(ns)
+
+        # Full GEE-native computation
+        result = ha.compute_hydroperiod(index='mndwi', threshold=0.1)
+
+        # result is an ee.Image with bands:
+        #   hydroperiod, valid_days, normalized,
+        #   first_flood_doy, last_flood_doy
+
+        # Export to Drive
+        ha.export_to_drive(folder='my_wetland')
+
+        # Per-pixel IRT (temporal representativity)
+        irt = ha.compute_irt_image()
+
+    Notes
+    -----
+    Supported water indices (optical sensors):
+    ``'ndwi'``, ``'mndwi'``, ``'awei'``, ``'aweinsh'``, ``'wi2015'``
+
+    The per-pixel IRT is computed as Simpson's diversity index across monthly
+    periods, which is fully computable per-pixel in GEE. The global IRT uses
+    the Gini-based formula identical to the standalone *phydroperiod* library
+    (computed client-side from scene dates).
+    """
+
+    #: Water indices supported for hydroperiod computation.
+    WATER_INDICES = frozenset({'ndwi', 'mndwi', 'awei', 'aweinsh', 'wi2015'})
+
+    def __init__(self, ndvi_seasonality, hydrological_year_start=(9, 1)):
+        self.ns = ndvi_seasonality
+        self.hyd_month, self.hyd_day = hydrological_year_start
+        self._water_masks = None
+        self._hyd_year = None
+        self._d0 = None
[…]
