### 2026-04-13 #9763 @mengqi-z
@Digdgeo - Thanks! I will update the version to v1.2.2. Usually, the version listed here will not affect the review process. We only need to make sure the final accepted package has the latest version.
