### 2026-04-15 48789d5
Update JOSS paper: add HydroperiodAnalyzer section and new workflow figure
- Add Hydroperiod analysis section to Key Features describing flood days,
  valid days, normalized flood days, timing metrics, anomalies, and
  configurable hydrological cycle boundaries
- Update Summary to mention hydroperiod module
- Move Wohner2019 citation from table to introductory text
- Replace workflow figure with updated version including Hydroperiod module
