+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-Google Earth Engine (GEE) is a cloud-based platform for planetary-scale geospatial analysis [@Gorelick2017]. Ndvi2Gif is a Python package built on top of GEE and geemap [@Wu2020] that simplifies the generation and analysis of multi-temporal composite images from satellite and climate reanalysis data. The package provides unified access to 7 satellite platforms (Sentinel-1/2/3, Landsat 4–9, MODIS) and climate reanalysis data (ERA5-Land, CHIRPS), with 52 predefined spectral and radar indices plus 48 climate variables.
+Google Earth Engine (GEE) is a cloud-based platform for planetary-scale geospatial analysis [@Gorelick2017]. Ndvi2Gif is a Python package built on top of GEE and geemap [@Wu2020] that simplifies the generation and analysis of multi-temporal composite images from satellite and climate reanalysis data. The package provides unified access to 7 satellite platforms (Sentinel-1/2/3, Landsat 4–9, MODIS) and climate reanalysis data (ERA5-Land, CHIRPS), with 52 predefined spectral and radar indices plus 48 climate variables. Beyond compositing, the package includes modules for SAR preprocessing, time series and phenology analysis, machine learning land cover classification, and hydroperiod analysis of surface water dynamics.
 
 The core concept of Ndvi2Gif is the creation of **multi-seasonal statistical composites**. The `get_year_composite()` method returns an `ee.ImageCollection` where each image represents one year, with bands for each temporal period (e.g., `periods=12` generates bands named `january`, `february`, ..., `december`). This band-based temporal organization enables queries such as: *What was the maximum flood extent detected in January across 40 years of Landsat (1984–2024)?* by simply calling `collection.select('january').max()`. Built-in zonal statistics extraction via `get_stats()` allows direct computation of statistics per polygon without external GIS processing.
 
@@ -72,7 +72,7 @@ Ndvi2Gif provides unified access to optical, SAR, and climate reanalysis dataset
 
 ## Flexible region of interest
 
-The library supports multiple ROI specification formats, enabling seamless integration with existing workflows and the eLTER research network.
+The library supports multiple ROI specification formats, enabling seamless integration with existing workflows and the eLTER research network [@Wohner2019].
 
 : Supported region of interest formats. \label{tab:roi}
 
@@ -80,7 +80,7 @@ The library supports multiple ROI specification formats, enabling seamless integ
 | :---------- | :---------------------- | :----------------------- |
 | Shapefile   | `'study_area.shp'`      | Local vector file        |
 | GeoJSON     | `'boundaries.geojson'`  | GeoJSON file             |
-| DEIMS ID    | `'deimsid:8eda49e9...'` | eLTER site [@Wohner2019] |
+| DEIMS ID    | `'deimsid:8eda49e9...'` | eLTER site               |
 | S2 tile     | `'s2:29SQB'`            | MGRS tile code           |
 | Landsat WRS | `'wrs:200,32'`          | Path and Row             |
 | ee.Geometry | `ee.Geometry.Point()`   | GEE geometry             |
@@ -116,6 +116,22 @@ This design allows researchers to move seamlessly from data extraction to explor
 The `LandCoverClassifier` module provides supervised and unsupervised land cover classification workflows based on multi-temporal composite stacks generated by `NdviSeasonality`. Supported supervised algorithms include Random Forest, Support Vector Machines (SVM), CART, and Gradient Tree Boost, while unsupervised approaches include K-means and Cascade K-means clustering.
 
[…]
