### 2026-04-15 8bf85a0
Add tutorials, Hydroperiod notebook, and fix docs config
- Add 5 tutorial MDs (ndvi2gif, timeseries, s1_ard, clasification, hydroperiod)
- Add Hydroperiod.ipynb example notebook to docs
- Integrate tutorials and Hydroperiod notebook into Jupyter Book toctree
- Fix GitHub URL in conf.py (was placeholder tuusuario/ndvi2gif)
