+++ docs/tutorials/ndvi2gif.md
@@ -1,0 +1,351 @@
+# Temporal Compositing with `NdviSeasonality`
+
+`NdviSeasonality` is the core engine of ndvi2gif. It configures the satellite collection, computes spectral and radar indices, and generates temporal composite images — one per period per year. All other classes (`HydroperiodAnalyzer`, `TimeSeriesAnalyzer`, `LandCoverClassifier`) take an existing `NdviSeasonality` instance and build on top of it.
+
+---
+
+## What are temporal composites?
+
+Rather than working with individual satellite scenes (which are cloudy, noisy, and irregularly spaced), temporal compositing aggregates all cloud-free observations within a time window into a single representative image using a statistical reducer (max, median, mean, percentile).
+
+A **seasonal composite** for NDVI answers: "what was the maximum NDVI in spring 2022 across every pixel?"
+
+ndvi2gif supports:
+
+- **4 periods/year** — meteorological seasons (winter, spring, summer, autumn)
+- **12 periods/year** — calendar months (january through december)
+- **24 periods/year** — bi-monthly (~15 days each)
+- **Any N** — equal divisions of the year (p1, p2, …, pN)
+
+---
+
+## Quick start
+
+```python
+import ee
+import geemap
+from ndvi2gif import NdviSeasonality
+
+ee.Initialize()
+
+roi = ee.Geometry.Rectangle([-6.55, 36.85, -6.10, 37.20])
+
+ns = NdviSeasonality(
+    roi=roi,
+    sat='S2',
+    start_year=2019,
+    end_year=2023,
+    periods=4,
+    index='ndvi',
+    key='max',
+)
+
+# Get composite ImageCollection (one image per year, each with 4 period bands)
+collection = ns.get_year_composite()
+print(collection.size().getInfo())  # → 5 (one per year)
+```
+
+---
+
+## Region of interest (ROI)
+
+The `roi` parameter accepts multiple formats:
+
+```python
+# Direct EE geometry
+roi = ee.Geometry.Rectangle([-6.55, 36.85, -6.10, 37.20])
+roi = ee.Geometry.Point([-5.97, 37.27]).buffer(5000)  # 5 km buffer
+
+# Shapefile or GeoJSON
+ns = NdviSeasonality(roi='study_area.shp')
+ns = NdviSeasonality(roi='study_area.geojson')
+
+# eLTER DEIMS site ID
+ns = NdviSeasonality(roi='deimsid/11696159-444f-4e06-b537-d4c5c0a4e97d')
+
+# Sentinel-2 MGRS tile
+ns = NdviSeasonality(roi='s2:30TXN')
+
+# Landsat WRS-2 path/row
+ns = NdviSeasonality(roi='wrs:200,32')
+
+# Drawn features from a geemap Map
+ns = NdviSeasonality(roi=Map.draw_features)
+
+# None → default region (Andalusia, Spain — useful for quick tests)
+ns = NdviSeasonality()
+```
+
+---
+
+## Satellite sensors
+
+| `sat=` | Sensor | Resolution | Archive start | Best for |
+|---|---|---|---|---|
+| `'S2'` | Sentinel-2 MSI | 10–20 m | 2015 | Detailed land cover, vegetation |
+| `'Landsat'` | Landsat 4–5–7–8–9 | 30 m | 1982 | Long time series |
+| `'MODIS'` | Terra/Aqua | 500 m | 2000 | Large-scale, daily coverage |
+| `'S1'` | Sentinel-1 SAR | 10 m | 2014 | All-weather, structural changes |
+| `'S3'` | Sentinel-3 OLCI | 300 m | 2016 | Ocean/coastal, water quality |
+
+```python
+# Landsat for long-term analysis (1990–2023)
[…]
