+++ docs/tutorials/timeseries.md
@@ -1,0 +1,320 @@
+# Time Series and Trend Analysis
+
+ndvi2gif provides two complementary classes for temporal analysis:
+
+- **`TimeSeriesAnalyzer`** — extracts time series for a point or region, runs statistical trend tests, and generates multi-panel analysis dashboards with phenological metrics.
+- **`SpatialTrendAnalyzer`** — computes per-pixel trend maps across the entire ROI using GEE's server-side linear regression.
+
+Both classes take a configured `NdviSeasonality` instance and inherit all its settings (ROI, satellite, date range, index).
+
+---
+
+## Quick start
+
+```python
+import ee
+import geemap
+from ndvi2gif import NdviSeasonality, TimeSeriesAnalyzer, SpatialTrendAnalyzer
+
+ee.Initialize()
+
+roi = ee.Geometry.Rectangle([-6.55, 36.85, -6.10, 37.20])
+
+ns = NdviSeasonality(
+    roi=roi,
+    sat='S2',
+    index='ndvi',
+    periods=12,          # monthly composites
+    start_year=2017,
+    end_year=2023,
+)
+
+analyzer = TimeSeriesAnalyzer(ns)
+df = analyzer.extract_time_series()
+print(df.head())
+```
+
+---
+
+## `TimeSeriesAnalyzer`
+
+### Extracting a time series
+
+`extract_time_series()` queries each temporal period across all years and returns a `pandas.DataFrame`.
+
+```python
+# Extract from ROI centroid (default)
+df = analyzer.extract_time_series()
+
+# Extract from a specific point (lon, lat)
+df = analyzer.extract_time_series(point=(-5.97, 37.27))
+
+# Extract mean value over the full ROI (polygon reduction)
+df = analyzer.extract_time_series(
+    point=roi,
+    reducer='mean',
+    scale=20,
+)
+```
+
+The returned DataFrame has columns: `date`, `value`, `year`, `period`, `doy` (day of year), `season`, `month`.
+
+> **Performance note:** `extract_time_series()` calls `reduceRegion` for each period of each year — it sends one GEE request per period. For a 6-year monthly analysis that is 72 requests. The results are cached automatically; subsequent calls with the same parameters return the cached DataFrame instantly.
+
+---
+
+### Trend analysis
+
+`analyze_trend()` runs one or more statistical tests on the extracted time series and returns a summary dictionary.
+
+```python
+# Mann-Kendall non-parametric test (recommended for environmental data)
+trends = analyzer.analyze_trend(method='mann_kendall')
+print(trends['interpretation'])
+# → "Significant increasing trend (p=0.012)"
+
+# All methods at once
+trends = analyzer.analyze_trend(method='all', alpha=0.05)
+
+# Mann-Kendall result
+mk = trends['mann_kendall']
+print(f"tau={mk['tau']:.3f}, p={mk['p_value']:.4f}, trend={'↑' if mk['trend'] > 0 else '↓'}")
+
+# Linear regression
+lr = trends['linear']
+print(f"slope={lr['slope']:.4f}/period, R²={lr['r_squared']:.3f}")
+print(f"Annual change: {lr['yearly_change']:.4f}")
+
+# Sen's slope (robust estimate)
+sen = trends['sen_slope']
+print(f"Median slope: {sen['median_slope']:.4f}")
+```
+
+Available methods:
+
+| `method=` | Test | Notes |
+|---|---|---|
+| `'mann_kendall'` | Mann-Kendall non-parametric test | Robust to non-normality, recommended for NDVI |
+| `'linear'` | Ordinary least squares regression | Assumes linearity and normality |
+| `'sen_slope'` | Theil-Sen estimator | Robust to outliers, median-based |
+| `'all'` | All three methods | |
+
+---
+
[…]
