### 2026-04-15 d81e3b6
Add HydroperiodAnalyzer to book: tutorial, notebook, and fixes
- Add book/advanced/hydroperiod.md with full HydroperiodAnalyzer tutorial
- Add book/notebooks/05_hydroperiod.ipynb example notebook
- Update _toc.yml to include hydroperiod in Advanced Features and Notebooks
- Fix installation.md: remove numpy<2.0 pin, add HydroperiodAnalyzer to imports, update version to 1.2.2
