+++ book/advanced/hydroperiod.md
@@ -1,0 +1,521 @@
+# Hydroperiod Analysis with `HydroperiodAnalyzer`
+
+`HydroperiodAnalyzer` computes **hydroperiod** — the number of days a pixel remains flooded during a hydrological cycle — entirely in Google Earth Engine. No data downloads are required: all computation runs server-side as a lazy GEE graph that is only evaluated when you add a layer to a map or trigger an export.
+
+---
+
+## What is hydroperiod?
+
+Hydroperiod is a fundamental ecological indicator for wetland ecosystems. It quantifies how long each pixel is covered by water during a defined hydrological cycle (by default September 1st – August 31st). It drives:
+
+- Plant community composition and distribution
+- Habitat suitability for waterbirds, amphibians, and invertebrates
+- Carbon dynamics and greenhouse gas fluxes
+- Long-term wetland monitoring and conservation planning
+
+---
+
+## Methodology: midpoint temporal weighting
+
+Raw satellite archives contain an uneven number of cloud-free observations over time. Simply counting flood detections ignores this irregularity — a pixel observed only in winter would appear drier than one observed year-round, even if both were equally flooded.
+
+The **midpoint weighting method** solves this by distributing the 365 days of the cycle proportionally among available scenes:
+
+1. Sort all cloud-free scenes by acquisition date within the cycle.
+2. For each consecutive pair of scenes, compute the temporal midpoint.
+3. Each scene is assigned a **territorial weight** spanning from the midpoint with its predecessor to the midpoint with its successor. The first and last scenes use day 0 and day 365 as outer boundaries.
+4. A pixel contributes `weight` flood days if it is water in that scene, and `weight` valid days regardless (as long as it is not cloud-masked).
+
+```
+Scenes (days from Sep 1):   0    14    45    120   230   310
+Midpoints:                      7    29    82    175   270
+Territories:             [0,7] [7,29] [29,82] [82,175] [175,270] [270,365]
+Weights (days):           7     22     53      93       95        95
+```
+
+The sum of all weights is always exactly 365. The final hydroperiod for a pixel is the sum of weights across scenes where that pixel was classified as water.
+
+### Same-day tile mosaicking
+
+Sentinel-2 and other sensors can produce several overlapping scenes from different tiles on the same calendar day. Without correction, same-day scenes get a midpoint distance of zero → weight of 0 days, silently discarding valid observations.
+
+`HydroperiodAnalyzer` solves this by mosaicking all scenes acquired on the same day with `.max()` before computing weights: if *any* tile shows water for a given pixel on that day, the pixel is classified as water.
+
+---
+
+## Installation and setup
+
+```python
+import ee
+import geemap
+from ndvi2gif import NdviSeasonality, HydroperiodAnalyzer
+
+# Authenticate only the first time
+# ee.Authenticate()
+ee.Initialize()
+```
+
+`HydroperiodAnalyzer` requires an existing `NdviSeasonality` instance. All satellite configuration — ROI, sensor, date range, cloud filtering, band standardisation — is reused from it.
+
+---
+
+## Quick start
+
+```python
+roi = ee.Geometry.Rectangle([-6.55, 36.85, -6.10, 37.20])  # Doñana, Spain
+
+ns = NdviSeasonality(
+    roi=roi,
+    sat='S2',
+    start_year=2022,
[…]
