### 2026-04-15 81f50b0
Switch book theme to sphinx_book_theme (white background, lateral menu)
- Replace pydata_sphinx_theme with sphinx_book_theme for cleaner look
- Fix numpy pin: <2.0 → >=1.24 in requirements.txt
