+++ book/requirements.txt
@@ -2,7 +2,7 @@
 jupyter-book>=0.15.1,<2.0  # Pin to 1.x to avoid EISDIR errors in 2.x
 ghp-import
 sphinx>=5.0
-pydata-sphinx-theme>=0.13.0
+sphinx-book-theme>=1.0.0
 sphinx-copybutton
 myst-nb>=0.17.0
 sphinxcontrib-bibtex
@@ -11,7 +11,7 @@ sphinxcontrib-bibtex
 ndvi2gif>=0.6.0
 earthengine-api>=0.1.347
 geemap>=0.29.5
-numpy<2.0
+numpy>=1.24
 pandas
 matplotlib
 seaborn
