+++ book/_config.yml
@@ -50,25 +50,15 @@ sphinx:
     - sphinx.ext.napoleon
     - sphinx.ext.intersphinx
   config:
-    html_theme: pydata_sphinx_theme
+    html_theme: sphinx_book_theme
     html_theme_options:
-      github_url: https://github.com/Digdgeo/Ndvi2Gif
       repository_url: https://github.com/Digdgeo/Ndvi2Gif
       repository_branch: master
       use_repository_button: true
       use_issues_button: true
       use_edit_page_button: true
-      show_toc_level: 2
-      navbar_align: left
-      logo:
-        text: Ndvi2Gif
-      icon_links:
-        - name: PyPI
-          url: https://pypi.org/project/ndvi2gif/
-          icon: fas fa-box
-        - name: Conda
-          url: https://anaconda.org/conda-forge/ndvi2gif
-          icon: fas fa-cloud
+      show_navbar_depth: 2
+      home_page_in_toc: true
     intersphinx_mapping:
       python:
         - https://docs.python.org/3
