### 2026-04-15 e1a95fe
Fix notebook titles in TOC and update citation year
- Add clean display titles to all notebooks (remove number prefixes)
- Fix citation: year 2020 (first release), version 1.2.2 (current)
