### 2026-04-15 e5d19b8
Replace notebook stubs with real notebooks, fix FAQ numpy entry
- Add Basic_Usage, LandCover, TimeSeries real notebooks to book
- Remove corresponding stub MD files
- Remove outdated numpy<2.0 FAQ entry
