+++ book/notebooks/02_landcover_classification.md
@@ -1,3 +1,0 @@
-# 02 landcover classification
-
-*Notebook coming soon. In the meantime, check the [examples_notebooks](https://github.com/Digdgeo/Ndvi2Gif/tree/master/examples_notebooks) in the repository.*
