+++ book/notebooks/01_basic_usage.md
@@ -1,3 +1,0 @@
-# 01 basic usage
-
-*Notebook coming soon. In the meantime, check the [examples_notebooks](https://github.com/Digdgeo/Ndvi2Gif/tree/master/examples_notebooks) in the repository.*
