+++ book/notebooks/03_time_series_analysis.md
@@ -1,3 +1,0 @@
-# 03 time series analysis
-
-*Notebook coming soon. In the meantime, check the [examples_notebooks](https://github.com/Digdgeo/Ndvi2Gif/tree/master/examples_notebooks) in the repository.*
