### 2026-04-15 04f41a1
Fix paper inconsistencies and update README citation version
- Fix Landsat coverage year: 1984 → 1982 in example query
- Fix Sentinel-3 variables: 10 water quality → 36 optical
- Fix typo in workflow figure (statistcis → statistics)
- Update README citation version to 1.2.2
