### 2026-04-15 daed593
Finalize paper: consistent index counts, updated figure, minor fixes
- Update index counts to 53 total (26 generic + 10 S2 + 10 S3 + 7 S1)
- Fix per-sensor counts in Table 1 (S2/S3: 36, Landsat/MODIS: 27)
- Add explanatory note below Table 1 for generic vs specific indices
- Update workflow figure with Generic/Specific labels and correct counts
- Add Naive Bayes to ML classification section
- Fix Landsat coverage year (1982) and Sentinel-3 index description
