+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-Google Earth Engine (GEE) is a cloud-based platform for planetary-scale geospatial analysis [@Gorelick2017]. Ndvi2Gif is a Python package built on top of GEE and geemap [@Wu2020] that simplifies the generation and analysis of multi-temporal composite images from satellite and climate reanalysis data. The package provides unified access to 7 satellite platforms (Sentinel-1/2/3, Landsat 4–9, MODIS) and climate reanalysis data (ERA5-Land, CHIRPS), with 52 predefined spectral and radar indices plus 48 climate variables. Beyond compositing, the package includes modules for SAR preprocessing, time series and phenology analysis, machine learning land cover classification, and hydroperiod analysis of surface water dynamics.
+Google Earth Engine (GEE) is a cloud-based platform for planetary-scale geospatial analysis [@Gorelick2017]. Ndvi2Gif is a Python package built on top of GEE and geemap [@Wu2020] that simplifies the generation and analysis of multi-temporal composite images from satellite and climate reanalysis data. The package provides unified access to 7 satellite platforms (Sentinel-1/2/3, Landsat 4–9, MODIS) and climate reanalysis data (ERA5-Land, CHIRPS), with 53 predefined spectral and radar indices plus 48 climate variables. Beyond compositing, the package includes modules for SAR preprocessing, time series and phenology analysis, machine learning land cover classification, and hydroperiod analysis of surface water dynamics.
 
 The core concept of Ndvi2Gif is the creation of **multi-seasonal statistical composites**. The `get_year_composite()` method returns an `ee.ImageCollection` where each image represents one year, with bands for each temporal period (e.g., `periods=12` generates bands named `january`, `february`, ..., `december`). This band-based temporal organization enables queries such as: *What was the maximum flood extent detected in January across 40 years of Landsat (1982–2024)?* by simply calling `collection.select('january').max()`. Built-in zonal statistics extraction via `get_stats()` allows direct computation of statistics per polygon without external GIS processing.
 
@@ -38,7 +38,7 @@ ndvi = NdviSeasonality(roi=roi, start_year=2013, end_year=2013,
 collection = ndvi.get_year_composite()
 ```
 
-This simplification extends across 7 satellite platforms, 52 spectral and radar indices, 48 climate variables, multiple temporal aggregation methods, and flexible region of interest specifications—all through a unified API that inherits geemap's interactive mapping capabilities. The architectural decisions and design philosophy that enable this simplification are detailed in the sections below. \autoref{fig:workflow} provides an overview of the package's workflow and capabilities.
+This simplification extends across 7 satellite platforms, 53 spectral and radar indices, 48 climate variables, multiple temporal aggregation methods, and flexible region of interest specifications—all through a unified API that inherits geemap's interactive mapping capabilities. The architectural decisions and design philosophy that enable this simplification are detailed in the sections below. \autoref{fig:workflow} provides an overview of the package's workflow and capabilities.
 
 ![Ndvi2Gif workflow architecture and capabilities. Color coding: Blue — input data sources and ROI specification; Green — processing operations; Purple — main analysis classes; Orange — download-ready output products.\label{fig:workflow}](ndvi2gif_workflow.png){ width=95% }
 
@@ -62,13 +62,15 @@ Ndvi2Gif provides unified access to optical, SAR, and climate reanalysis dataset
 
 | Platform   | Collection       | Resolution | Coverage     | Variables        |
 | :--------- | :--------------- | :--------- | :----------- | :--------------- |
[…]
