### 2026-04-15 #9763 @Digdgeo
Hi @mengqi-z, I just wanted to point out that I updated the paper today to include a new section covering the HydroperiodAnalyzer module, introduced in v1.2.2. Since the review hasn't started yet, I hope this is fine!
Thanks a lot
