### 2026-04-16 d416a43
Add water quality tutorial and fix notebook rendering
Rewrite book/advanced/water_quality.md as a full advanced tutorial
based on Water_Quality_Embalses.ipynb (water detection with MNDWI,
monthly flooded surface, NDCI chlorophyll-a proxy masked over water,
trend analysis with TimeSeriesAnalyzer, exports).

Strip malformed metadata.widgets from the notebook so GitHub and
nbviewer render it instead of raising "Invalid Notebook" (the
widget-state dict was missing the required 'state' key).
