+++ book/advanced/water_quality.md
@@ -1,5 +1,320 @@
-# Water_quality
+# Water Quality Monitoring in Reservoirs
 
-*This section is under development and will be available soon.*
+This tutorial shows how to use **ndvi2gif** to monitor inland water bodies with Sentinel-2: detect the flooded extent each month, compute a monthly chlorophyll-a proxy, and study its inter-annual evolution. The workflow relies on the standard `NdviSeasonality` composites and on `TimeSeriesAnalyzer` for trend detection.
 
-Check back later or contribute via [GitHub](https://github.com/Digdgeo/Ndvi2Gif).
+A complete, runnable companion notebook with a reservoirs shapefile (CHG — Confederación Hidrográfica del Guadalquivir) is available in [`examples_notebooks/Water_Quality_Embalses.ipynb`](https://github.com/Digdgeo/Ndvi2Gif/blob/master/examples_notebooks/Water_Quality_Embalses.ipynb).
+
+---
+
+## Overview
+
+Reservoirs are heterogeneous systems: shorelines move with the water level, turbidity varies seasonally, and eutrophication episodes (cyanobacterial blooms) are concentrated in the warmer months. Any water-quality analysis therefore needs two layers:
+
+1. **Water detection** — a binary mask of "where is water *right now*".
+2. **Quality indicator** — a biogeochemical proxy evaluated **only on water pixels** of that same period.
+
+`ndvi2gif` makes this straightforward: water indices (NDWI, MNDWI, AWEI, WI2015) and the Red-Edge chlorophyll index NDCI are all available out of the box, all using the same standardised band names.
+
+| Index | Formula (S2) | Role |
+|---|---|---|
+| `ndwi` | `(Green − NIR) / (Green + NIR)` | Classic water detection |
+| `mndwi` | `(Green − SWIR1) / (Green + SWIR1)` | **Default water mask** — more robust |
+| `awei` / `aweinsh` | Multi-band (Blue, Green, NIR, SWIR1, SWIR2) | Automatic water extraction |
+| `wi2015` | NIR + SWIR1 + SWIR2 linear combination | Complementary water index |
+| `ndci` | `(RedEdge1 − Red) / (RedEdge1 + Red)` | **Chlorophyll-a proxy — Sentinel-2 only** |
+
+NDCI exploits the Red-Edge band B5 (705 nm), exclusive to Sentinel-2 among the sensors ndvi2gif supports, and is sensitive to chlorophyll-a concentration and potential cyanobacterial blooms.
+
+---
+
+## 1. Setup and ROI
+
+The ROI is a reservoir polygon. The notebook loads a CHG reservoir shapefile and selects one reservoir; any shapefile or GeoJSON works. A 500 m buffer is recommended so that the shoreline and the transition zone are included in every monthly composite, regardless of the water level:
+
+```python
+import ee, geemap, geopandas as gpd
+from ndvi2gif import NdviSeasonality, TimeSeriesAnalyzer
+
+ee.Initialize()
+
+gdf = gpd.read_file('CHG.ZonasProtegidas.CaptacionesEmbalses.shp')
+gdf['area_m2'] = gdf.geometry.to_crs('EPSG:32630').area
+
+# Pick the largest reservoir and buffer it by 500 m
+reservoir = gdf.loc[[gdf['area_m2'].idxmax()]]
+roi_buffer = reservoir.to_crs('EPSG:32630').buffer(500).to_crs('EPSG:4326')
+roi_path = '/tmp/reservoir_buffer.shp'
+gpd.GeoDataFrame(geometry=roi_buffer, crs='EPSG:4326').to_file(roi_path)
+```
+
[…]
