### 2026-04-16 2570ffc
Add SAR processing advanced tutorial
Replace stub book/advanced/sar_processing.md with a full tutorial on
S1ARDProcessor: speckle filters, radiometric terrain correction, DEM
and orbit choices, the seven SAR indices available through
NdviSeasonality, a crop-phenology walk-through, and an API reference.
