+++ book/advanced/sar_processing.md
@@ -1,5 +1,340 @@
-# Sar_processing
+# Sentinel-1 SAR Processing with `S1ARDProcessor`
 
-*This section is under development and will be available soon.*
+Sentinel-1 C-band SAR sees through clouds and night, and therefore fills the gap left by optical sensors in cloud-prone areas (boreal, tropical, monsoon) and in rapidly changing systems (crops, wetlands, floods). But raw GRD scenes are not directly comparable in time: they are affected by speckle noise and by topographic distortions that depend on the scene's acquisition geometry.
 
-Check back later or contribute via [GitHub](https://github.com/Digdgeo/Ndvi2Gif).
+`S1ARDProcessor` turns Sentinel-1 GRD data into **Analysis Ready Data (ARD)** inside Google Earth Engine. It applies radiometric terrain correction and one of five speckle filters, producing time-consistent VV and VH backscatter that can be used for indices, composites, classification, or time-series analysis.
+
+The processor follows the angular-based radiometric slope correction method of Vollrath et al. (2020).
+
+---
+
+## What ARD gives you
+
+| Problem | Consequence | What ARD does |
+|---|---|---|
+| Speckle noise | High pixel-to-pixel variance, unreliable thresholds | Adaptive spatial filtering (Lee variants, Gamma MAP, Lee Sigma, Boxcar) |
+| Terrain distortion | Same land cover with different backscatter on pole-facing vs sun-facing slopes | Radiometric correction from local incidence angle (slope, aspect, orbit) |
+| Geometric inconsistency between orbits | Asc vs desc scenes not comparable | Filter by `orbit` keeping only one pass direction |
+| Raw linear power is hard to interpret | Non-Gaussian dynamic range | Optional conversion to decibel scale |
+
+In short: without ARD, pixel values between two dates are not directly comparable. With ARD, they are.
+
+---
+
+## Two ways to use it
+
+`S1ARDProcessor` can be used **directly** on any `ee.ImageCollection` of Sentinel-1 GRD scenes, or indirectly **through `NdviSeasonality`** by passing `use_sar_ard=True`. The latter is the usual path: it reuses the same ROI, date range, and orbit logic that you already configure for optical sensors.
+
+```python
+import ee
+from ndvi2gif import NdviSeasonality, S1ARDProcessor
+
+ee.Initialize()
+```
+
+---
+
+## Quick start — through `NdviSeasonality`
+
+```python
+roi = ee.Geometry.Rectangle([-6.55, 36.85, -6.10, 37.20])
+
+ns = NdviSeasonality(
+    roi=roi,
+    sat='S1',
+    index='vh',
+    start_year=2022,
+    end_year=2023,
+    periods=12,
+    key='median',
+    orbit='DESCENDING',       # pick one pass direction for temporal consistency
+    use_sar_ard=True,          # enable ARD pipeline
+    sar_speckle_filter='REFINED_LEE',
+    sar_terrain_correction=True,
+    sar_terrain_model='VOLUME',
+)
+
+composite = ns.get_year_composite()
+# ee.ImageCollection: 2 images (2022, 2023), each with 12 bands (january..december)
+```
+
+That's the full workflow: orbit-filtered, terrain-corrected, speckle-filtered, reduced to monthly medians of VH backscatter. Everything happens lazily in GEE — no download.
+
+---
+
+## Direct use — any GRD collection
+
[…]
