### 2026-04-16 207e49c
Fill advanced tutorials and add phydroperiod reference
- Write full tutorials for time series, classification and export options
- Add phydroperiod library pointer to hydroperiod notebook and tutorial
