+++ book/advanced/classification.md
@@ -1,5 +1,321 @@
-# Classification
+# Land Cover Classification with `LandCoverClassifier`
 
-*This section is under development and will be available soon.*
+`LandCoverClassifier` is the machine-learning layer on top of `NdviSeasonality`. It takes the multi-temporal composite you already configured — any sensor, any index, any period scheme — turns it into a feature stack, and feeds it into one of the supervised or unsupervised classifiers available in Google Earth Engine. Training, prediction, and accuracy assessment all run server-side.
 
-Check back later or contribute via [GitHub](https://github.com/Digdgeo/Ndvi2Gif).
+The design intent: you should not have to rewrite your feature engineering every time you want to try a new classifier. The `NdviSeasonality` configuration *is* the feature engineering.
+
+---
+
+## Quick start
+
+```python
+import ee
+from ndvi2gif import NdviSeasonality, LandCoverClassifier
+
+ee.Initialize()
+
+ns = NdviSeasonality(
+    roi=roi,
+    sat='S2',
+    index='ndvi',
+    start_year=2021, end_year=2023,
+    periods=4,
+    key='median',
+)
+
+clf = LandCoverClassifier(ns)
+
+clf.create_feature_stack(indices=['ndvi', 'ndwi', 'bsi'], include_statistics=True)
+clf.add_training_data(training_polygons='training.shp', class_property='class')
+
+classified = clf.classify_supervised(algorithm='random_forest')
+
+print(clf.accuracy_results['overall_accuracy'], clf.accuracy_results['kappa'])
+```
+
+---
+
+## 1. Building a feature stack
+
+`create_feature_stack()` stacks every `period × year × index` combination into a single multi-band `ee.Image`. A three-year Sentinel-2 seasonal composite with three indices produces 3 × 4 × 3 = 36 bands.
+
+```python
+stack = clf.create_feature_stack(
+    indices=['ndvi', 'ndwi', 'bsi'],
+    include_statistics=True,   # adds mean, std, max, min per index
+    normalize=True,            # rescales all bands to [0, 1]
+)
+```
+
+| Argument | Purpose |
+|---|---|
+| `indices` | Indices to stack. If `None`, uses the current `NdviSeasonality.index`. Must be compatible with the sensor. |
+| `include_statistics` | Adds four temporal summary bands per index (`mean`, `std`, `max`, `min`) — often the most informative features for vegetation classes. |
+| `normalize` | Linear min-max rescaling to `[0, 1]`. Recommended for SVM and Naive Bayes, optional for tree-based methods. |
+
+Band names follow the pattern `{index}_{year}_{period}` — e.g. `ndvi_2021_summer`, `ndwi_2022_winter`, `bsi_mean`.
+
+> **Why multi-temporal helps.** A cropland pixel and a shrubland pixel can have identical summer NDVI. What separates them is the seasonal *trajectory* — the winter trough, the spring rise, the post-harvest drop. Stacking the periods makes that trajectory available to the classifier as features, instead of collapsing it into a single statistic.
+
+### Sensor compatibility
+
+The stack inherits the sensor from `NdviSeasonality`. Invalid indices (not registered for that sensor in `self.sensor_indices`) raise a `ValueError` before any GEE work runs.
+
+---
+
+## 2. Training data
+
+Two input formats are supported: **points** (already labelled) or **polygons** (a class label per polygon, from which points are sampled).
+
+```python
+# Option A — labelled points
+clf.add_training_data(
+    training_points='training_points.shp',
+    class_property='class',
+)
+
+# Option B — labelled polygons, sample 100 points each
+clf.add_training_data(
+    training_polygons='training_polygons.geojson',
[…]
