+++ book/advanced/export_options.md
@@ -1,5 +1,221 @@
-# Export_options
+# Export & Batch Processing
 
-*This section is under development and will be available soon.*
+`NdviSeasonality` offers several export paths, each aimed at a different scale of analysis. The guidance below is opinionated: pick the smallest tool that still fits your job.
 
-Check back later or contribute via [GitHub](https://github.com/Digdgeo/Ndvi2Gif).
+| You want… | Use this |
+|---|---|
+| A single GeoTIFF on your laptop, ROI < ~2 500 km² | `get_export_single()` |
+| One GeoTIFF per year on your laptop | `get_export()` |
+| A large file you can leave running | `export_to_drive()` |
+| A reusable layer other GEE scripts can read | `export_to_asset()` |
+| An ROI too big for any single export | `export_with_fishnet()` |
+| An animated preview | `get_gif()` |
+
+All of these sit on the same Earth Engine export primitives (`Export.image.toDrive`, `Export.image.toAsset`). The wrappers exist to fill in ROI, scale, and sensible defaults from the `NdviSeasonality` configuration.
+
+---
+
+## 1. Local exports through `geemap`
+
+Local exports stream the image from GEE to your disk as a single GeoTIFF. They are convenient but have hard limits: roughly 2 GB per file and a few minutes of server time. If you routinely hit those limits, switch to Drive or Asset exports.
+
+### `get_export_single()` — one image
+
+```python
+image = ns.get_year_composite().first()
+ns.get_export_single(image, name='ndvi_2022.tif', scale=10, crs='EPSG:4326')
+```
+
+| Argument | Default | Description |
+|---|---|---|
+| `image` | — | Any `ee.Image`. Usually a composite from `get_year_composite()`. |
+| `name` | `'mycomposition.tif'` | File name written to the current working directory. |
+| `crs` | `'EPSG:4326'` | Output projection. Use a local UTM zone if you care about metric measurements on the result. |
+| `scale` | `10` | Pixel size in metres. **Match the sensor**, not a number you like: 10 for S2/S1, 30 for Landsat, 250–500 for MODIS, 300 for S3. |
+
+### `get_export()` — one GeoTIFF per year
+
+```python
+ns.get_export(scale=10, crs='EPSG:4326')
+# ndvi_max_2020.tif, ndvi_max_2021.tif, ndvi_max_2022.tif, ...
+```
+
+Filenames follow the pattern `{sat}_{index}_{stat}_{year}.tif`, where `{stat}` is `max`, `mean`, `median`, or `p{N}` when `key='percentile'`. One multi-band file per year; bands are the periods (`winter, spring, summer, autumn` for `periods=4`; `january…december` for `periods=12`; `p1…pN` otherwise).
+
+> **Don't oversample.** Setting `scale=10` on Landsat or MODIS data does not produce 10 m imagery — it produces a resampled fake that wastes quota. Use the native resolution of the sensor.
+
+---
+
+## 2. Drive exports — `export_to_drive()`
+
+For anything that won't fit comfortably into a local download, go through Drive. This starts an Earth Engine batch task, returns immediately, and writes the result to your Google Drive when the server is ready.
+
+```python
+task = ns.export_to_drive(
+    image=classified,
+    description='landcover_2022_RF',
+    folder='ndvi2gif_exports',
+    scale=10,
+    crs='EPSG:4326',
+    file_format='GeoTIFF',
+    format_options={'cloudOptimized': True, 'compression': 'LZW'},
+    max_pixels=int(1e13),
+    file_dimensions=10000,
+    clip_region=True,
+)
+```
+
+| Argument | Purpose |
+|---|---|
+| `description` | Task name shown in the Tasks panel. Also used as the base filename. |
[…]
