### 2026-04-29 #9763 @mengqi-z
👋 @davemlz @samvanstroud - Would any of you be interested and available to review this submission for the Journal of Open Source Software (JOSS)?
JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
