### 2026-04-30 #9763 @albhasan
> 👋 [@robbibt](https://github.com/robbibt) [@albhasan](https://github.com/albhasan) [@leouieda](https://github.com/leouieda) - Would any of you be interested and available to review this submission for the Journal of Open Source Software (JOSS)? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi, thanks for your kind invitation, but I don't think I can be of any help because I don't know much about GEE; I don't use it at all.
