### 2026-06-06 #10654 @editorialbot
[REVIEW]: Ndvi2Gif: A Python Package for Multi-Seasonal Remote Sensing Analysis via Google Earth Engine
**Submitting author:** @Digdgeo (<a href="http://orcid.org/0000-0002-2757-7112">Diego  García Díaz</a>)
**Repository:** https://github.com/Digdgeo/Ndvi2Gif
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.1
