### 2026-06-14 b3ddb64
Add SpatialPhenologyAnalyzer and release v1.3.0
New server-side, per-pixel phenology module producing SOS/POS/EOS rasters
across the whole ROI, with threshold, derivative and harmonic (Fourier)
methods. The harmonic method is the Earth Engine-native replacement for
the client-side double-logistic fit. Provides per-year (extract_phenology_rasters)
and multi-year aggregate (phenology_summary) outputs, and export to local
GeoTIFF (band names via rasterio) or Google Drive.

Bump version to 1.3.0; update CHANGELOG, README and add rasterio dependency.
