+++ ndvi2gif/timeseries.py
@@ -3251,5 +3251,509 @@ class SpatialTrendAnalyzer:
             filename = f'{self.processor.sat}_{self.processor.index}_trend_{method}_{self.processor.start_year}_{self.processor.end_year-1}.tif'
             print(f"Exporting trend map as {filename}")
             self.processor.get_export_single(trend_image, name=filename, scale=scale)
-        
+
         return trend_image
+
+
+class SpatialPhenologyAnalyzer:
+    """
+    Server-side, per-pixel phenology raster generation with Earth Engine.
+
+    While :class:`TimeSeriesAnalyzer` extracts phenological metrics for a single
+    point (downloading the time series via ``getInfo``), this class computes the
+    same family of metrics for **every pixel** of the ROI entirely on Earth
+    Engine, producing phenology rasters (Start/Peak/End of Season, etc.).
+
+    Three extraction methods are available, all running fully server-side:
+
+    * ``'threshold'`` : amplitude-based thresholding (SOS/EOS as the first/last
+      day-of-year above an amplitude fraction of the seasonal curve).
+    * ``'derivative'`` : SOS/EOS from the steepest positive/negative rate of
+      change between consecutive composites.
+    * ``'harmonic'`` : harmonic (Fourier) regression per pixel, used to
+      reconstruct a smooth seasonal curve from which SOS/POS/EOS are extracted.
+      This is the Earth Engine-friendly replacement for the client-side
+      double-logistic fit (which relies on ``scipy.optimize.curve_fit`` and
+      cannot run server-side).
+
+    Phenology is computed **year by year** (each year has its own seasonal
+    cycle). Two entry points are provided:
+
+    * :meth:`extract_phenology_rasters` returns one image per year as an
+      ``ee.ImageCollection`` (useful for analysing how phenology shifts over
+      time, e.g. earlier SOS year after year).
+    * :meth:`phenology_summary` collapses all years into a single multi-band
+      ``ee.Image`` using a reducer (median/mean), i.e. the typical phenology of
+      the period.
+
+    Output bands (per year and in the summary): ``sos``, ``pos``, ``eos``,
+    ``los``, ``amplitude``, ``peak_value``, ``baseline``, ``growth_rate``,
+    ``senescence_rate``. SOS/POS/EOS/LOS are expressed in day-of-year (days).
+
+    Parameters
+    ----------
+    ndvi_seasonality_instance : NdviSeasonality
+        Configured NdviSeasonality instance with ROI, periods and year range.
+
+    Notes
+    -----
+    Phenology extraction needs enough intra-annual temporal resolution. Using
+    ``periods=12`` (monthly) or ``periods=24`` (bi-monthly) is strongly
+    recommended; with ``periods=4`` (seasonal) results are coarse and a warning
+    is issued.
+
+    Examples
+    --------
+    >>> processor = NdviSeasonality(sat='S2', index='ndvi', periods=12,
+    ...                             start_year=2018, end_year=2023)
+    >>> pheno = SpatialPhenologyAnalyzer(processor)
+    >>> # Option A: one phenology image per year
+    >>> yearly = pheno.extract_phenology_rasters(method='harmonic')
+    >>> # Option B: single multi-year median phenology map
+    >>> summary = pheno.phenology_summary(method='harmonic', reducer='median')
+    """
+
+    def __init__(self, ndvi_seasonality_instance):
+        """Initialize SpatialPhenologyAnalyzer with an NdviSeasonality instance."""
+        self.processor = ndvi_seasonality_instance
+
+    # ========== PUBLIC API ==========
+
+    def extract_phenology_rasters(self,
+                                  method: str = 'threshold',
+                                  threshold_percentile: float = 50,
+                                  adaptive_threshold: bool = True,
[…]
