+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ndvi2gif
-version = 1.2.2
+version = 1.3.0
 description = Multi-seasonal remote sensing analysis suite: composites, SAR preprocessing, time series analytics, and land cover classification with Google Earth Engine
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -44,6 +44,7 @@ install_requires =
     requests
     geopandas
     fiona
+    rasterio
     pycrs
     deims>=4.0
     statsmodels>=0.13
