### 2026-06-14 9cfb9de
Document spatial phenology in book and JOSS paper
Add a 'Per-pixel spatial phenology' section to advanced/time_series.md
(SpatialPhenologyAnalyzer: methods, outputs, export, API reference) and a
paragraph to the paper describing the per-pixel harmonic phenology module.
