+++ paper/paper.md
@@ -113,6 +113,8 @@ The module supports:
 
 This design allows researchers to move seamlessly from data extraction to exploratory and quantitative temporal analysis within a single, reproducible workflow, and is particularly suited to long-term ecological monitoring, phenology validation, and climate–vegetation interaction studies.
 
+Complementing this point- and region-based analysis, the `SpatialPhenologyAnalyzer` module computes per-pixel phenological rasters (SOS, POS, EOS, season length, amplitude, and growth and senescence rates) across the entire region of interest, running entirely server-side on Google Earth Engine. In addition to threshold- and derivative-based extraction, it provides a per-pixel harmonic (Fourier) regression that reconstructs a smooth seasonal trajectory, offering a scalable, Earth Engine-native alternative to double-logistic curve fitting. Results can be returned as one map per year or as a multi-year aggregate, enabling spatially explicit phenology mapping and validation over large areas.
+
 ## Machine learning classification
 
 The `LandCoverClassifier` module provides supervised and unsupervised land cover classification workflows based on multi-temporal composite stacks generated by `NdviSeasonality`. Supported supervised algorithms include Random Forest, Support Vector Machines (SVM), CART, Naive Bayes, and Gradient Tree Boost, while unsupervised approaches include K-means and Cascade K-means clustering.
