+++ book/advanced/time_series.md
@@ -2,7 +2,7 @@
 
 `TimeSeriesAnalyzer` turns the period-by-period composites produced by `NdviSeasonality` into a full time-series analysis pipeline: point or polygon extraction, trend detection, phenology metrics, and diagnostic dashboards. Everything runs as a thin Python layer on top of GEE — extraction uses `reduceRegion` lazily, and only the resulting table is brought back to the client for statistics and plotting.
 
-For **per-pixel** trend maps over an ROI (rather than a single point or polygon), ndvi2gif also ships `SpatialTrendAnalyzer`, covered at the end of this page.
+For **per-pixel** maps over an ROI (rather than a single point or polygon), ndvi2gif also ships `SpatialTrendAnalyzer` (trend maps) and `SpatialPhenologyAnalyzer` (phenology rasters), both covered at the end of this page.
 
 ---
 
@@ -212,6 +212,106 @@ This runs entirely on GEE servers — no client-side loop over pixels, and it sc
 
 ---
 
+## 6. Per-pixel spatial phenology with `SpatialPhenologyAnalyzer`
+
+`TimeSeriesAnalyzer.extract_phenology_metrics()` gives you SOS/POS/EOS for **one point** (it downloads the series with `getInfo` and runs SciPy on the client). When you want **phenology rasters** — a per-pixel map of SOS, POS, EOS and derived metrics across the whole ROI — use `SpatialPhenologyAnalyzer`, which computes the same family of metrics entirely server-side.
+
+```python
+from ndvi2gif import NdviSeasonality, SpatialPhenologyAnalyzer
+
+ns = NdviSeasonality(
+    roi=roi, sat='S2', index='ndvi',
+    start_year=2019, end_year=2022,
+    periods=12,            # monthly — enough intra-annual resolution
+    key='median',          # per-period reducer (median/max/percentile/...)
+)
+
+pheno = SpatialPhenologyAnalyzer(ns)
+```
+
+### Two outputs
+
+```python
+# Option A — one phenology image per year (ee.ImageCollection, each tagged 'year')
+yearly = pheno.extract_phenology_rasters(method='harmonic')
+
+# Option B — a single multi-year aggregate (ee.Image), the typical phenology
+summary = pheno.phenology_summary(method='harmonic', reducer='median')
+```
+
+Both return images (or collections) with nine bands:
+
+| Band | Units | Meaning |
+|---|---|---|
+| `sos`, `pos`, `eos` | day of year | Start / Peak / End of Season |
+| `los` | days | Length of Season (`eos − sos`) |
+| `amplitude` | index value | `peak − baseline` |
+| `peak_value`, `baseline` | index value | Seasonal max / min |
+| `growth_rate`, `senescence_rate` | value/day | Approximate green-up / senescence rates |
+
+Option A is for studying **how phenology shifts over time** (e.g. an advancing SOS year after year — feed `yearly` to `SpatialTrendAnalyzer` for a trend map of the timing). Option B is a clean, stable **reference phenology map** of the area.
+
+### Three server-side methods
+
+| Method | Definition of SOS/EOS | Notes |
+|---|---|---|
+| `threshold` | First / last day-of-year above an amplitude fraction of the seasonal curve | Fast and robust; mirrors the point-based threshold method |
+| `derivative` | Day of the steepest increase / decrease between consecutive composites | Captures fast green-up; sensitive to temporal resolution |
+| `harmonic` | Per-pixel **Fourier** regression → smooth reconstructed curve → threshold extraction | Earth Engine-native replacement for the double-logistic fit |
+
[…]
