### 2026-06-14 ff69c3d
Cite geemap, eemont and wxee in the paper's state of the field
