### 2026-06-17 Digdgeo/Ndvi2Gif#2 @davemlz
JOSS Review: Suggested (small) change to Figure 1 in the submitted paper
Hi @Digdgeo!

I'm currently reviewing your paper in JOSS (https://github.com/openjournals/joss-reviews/issues/10654) and for now I have a small suggestion:

This box "Analysis Capabilities" uses the same linetype as the arrows in the whole figure and creates a clash. I would suggest using another linetype either for the box of for the arrows (e.g., `--`) ;)

Cheers!

David
