### 2026-06-17 Digdgeo/Ndvi2Gif#2 @Digdgeo
Hi @davemlz

Thank you for pointing this out. You are right that, in the previous version, the solid Analysis Capabilities border could be mistaken with a flow arrow, especially since several output arrows crossed near it. I have addressed this by improving both the layout and the styling  of the diagram:

- Layout simplification: The five analysis capabilities (Time Series Analysis, Spatial Phenology, Machine Learning, Hydroperiod, and Zonal Statistics) are now arranged in a single, uniform row, reflecting that they operate as peer capabilities on the same  ee.ImageCollection. This was a bit of a pity, as I had originally wanted to indicate that Spatial Phenology is related to Time Series Analysis; but since they sit at the same level within Ndvi2Gif, I think this layout is more accurate.

- Following the same simplification idea: The final outputs (Maps, GIFs, Charts, Classification, and Statistics) no longer branch from individual module boxes. They now emerge from a single output bus below the analysis layer, eliminating the crossing arrows that caused the visual confusion.

- Visual styling: The Analysis Capabilities container border has also been changed from solid to dashed, as you suggested. This clearly distinguishes it as a logical grouping rather than a functional flow line or arrow.

I have already pushed the new figure to the repository and asked the bot to regenerate the PDF.

Thanks again! :)
