### 2026-06-23 Digdgeo/Ndvi2Gif#2 @davemlz
Thank you for the revision!

But now:

I would say this is actually not that clear. I would put them all inside a box, right now they just look like a kind of interconnected network that is not completely understandable.
