### 2026-06-25 Digdgeo/Ndvi2Gif#3 @davemlz
JOSS Review: Changes in the documentation
Hi @Digdgeo!

This issue is related to your current submission to JOSS (https://github.com/openjournals/joss-reviews/issues/10654).

> [!NOTE]
> This Issue will be constantly updated until all the documentation code have been tested, therefore, new items will appear in the lists of to-dos if necessary. **Update**: I ran all the code until the end of the **Getting Started** section. I see that for the remaining sections I will also get errors. Given this, I will wait until the docs page is updated to continue with the review of the docs.

First of all, your docs for `ndvi2gif` are pretty comprehensive and amazing!

And since the documentation relies strongly on the webpage docs, I would suggest making the following modifications:

- [x] Tutorial yet to be completed: https://digdgeo.github.io/Ndvi2Gif/tutorials/multi_sensor.html
- [x] Tutorial not available: https://digdgeo.github.io/Ndvi2Gif/tutorials/temporal_composites.html
- [x] Tutorial not available: https://digdgeo.github.io/Ndvi2Gif/tutorials/statistical_methods.html
- [x] Tutorial not available: https://digdgeo.github.io/Ndvi2Gif/tutorials/indices_overview.html
- [x] Use case not available: https://digdgeo.github.io/Ndvi2Gif/use_cases/agriculture.html
- [x] Use case not available: https://digdgeo.github.io/Ndvi2Gif/use_cases/wetlands.html
- [x] Use case not available: https://digdgeo.github.io/Ndvi2Gif/use_cases/drought.html
- [x] Use case not available: https://digdgeo.github.io/Ndvi2Gif/use_cases/phenology.html
- [x] Notebook not available: https://digdgeo.github.io/Ndvi2Gif/notebooks/04_sar_vegetation.html
- [x] I recommend adding the complete references to the indices in https://digdgeo.github.io/Ndvi2Gif/reference/indices.html (you can extract the references from https://github.com/awesome-spectral-indices/awesome-spectral-indices)
- [x] Add the documentation URL to the About section of the repository.
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/roi_options.html#elter-site-deims-id works but it shows no available data for the specified configuration.

Nevertheless, there are some changes that I consider crucial for the documentation:

- [x] Match methods available in the repository with the methods available in the webpage (for instance, the `NdviSeasonality` class has a long list of methods in the repo that are not included in the API reference: https://digdgeo.github.io/Ndvi2Gif/reference/api.html). This is also happening with the `S1ARDProcessor`, `TimeSeriesAnalyzer`, and `LandCoverClassifier` classes.
- [x] Add the classes that are available in the repository to the API reference: https://digdgeo.github.io/Ndvi2Gif/reference/api.html (for instance, the `SpatialTrendAnalyzer` and the `SpatialPhenologyAnalyzer` classes are not included in the API reference)
- [x] Changelog not available in the webpage: https://digdgeo.github.io/Ndvi2Gif/reference/changelog.html
- [x] Contributing Guidelines not available in the webpage: https://digdgeo.github.io/Ndvi2Gif/reference/contributing.html
- [x] For the API Reference consider adding an easier way for the user to find methods, since there is no panel for navigating the classes beyond `Core Classes` and `Utility Functions`.
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/authentication.html#testing-your-authentication does not work: `EEException: Image.load: Image asset 'COPERNICUS/S2_SR_HARMONIZED/20230601T110621_20230601T110616_T30TVK' not found (does not exist or caller does not have access)`. I have tried several other images with no trouble.
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#step-4-get-the-composite does not work: `TypeError: NdviSeasonality.get_year_composite() got an unexpected keyword argument 'year'`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#step-5-visualize does not work: `EEException: Image.visualize: Cannot provide a palette when visualizing more than one band.`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#step-6-create-an-animated-gif does not work: `TypeError: NdviSeasonality.get_gif() got an unexpected keyword argument 'fps'`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#complete-example does not work: `TypeError: NdviSeasonality.get_year_composite() got an unexpected keyword argument 'year'`. Note that the `fps` arg will probably cause trouble too.
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#export-as-geotiff does not work: `TypeError: NdviSeasonality.get_export() got an unexpected keyword argument 'year'`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/roi_options.html#sentinel-2-tile does not work: `EEException: Invalid GeoJSON geometry.`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/roi_options.html#landsat-path-row does not work: `EEException: Invalid GeoJSON geometry.`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/roi_options.html#working-with-large-areas does not work: `EEException: Invalid GeoJSON geometry.`
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/roi_options.html#example-multiple-roi-approaches does not work: `TypeError: NdviSeasonality.get_year_composite() got an unexpected keyword argument 'year'`. ROI using tiles wouldn't work either.

New errors (as of 2026-07-04):

- [x] The code under https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#export-as-geotiff does not work: `Generating URL ... An error occurred while downloading. Total request size (297571740 bytes) must be less than or equal to 50331648 bytes.` This after running the previous snippet "Complete Example".
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/advanced/time_series.html#quick-start does not work: `NameError: name 'roi' is not defined`.
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/advanced/classification.html#quick-start does not work: `NameError: name 'roi' is not defined`.
- [x] The code under https://digdgeo.github.io/Ndvi2Gif/advanced/water_quality.html#setup-and-roi does not work: `DataSourceError: CHG.ZonasProtegidas.CaptacionesEmbalses.shp: No such file or directory`. Maybe adding the shapefile to the data sources in the package? Or a way to download the file?
- [x] The page https://digdgeo.github.io/Ndvi2Gif/notebooks/02_landcover_classification.html is partially written in Spanish.

David
