### 2026-06-25 Digdgeo/Ndvi2Gif#4 @davemlz
JOSS Review: in-line print status for `NdviSeasonality`
Hi @Digdgeo!

This issue is related to your current submission to JOSS (https://github.com/openjournals/joss-reviews/issues/10654).

When using `NdviSeasonality`:

I notice that the status also prints the processing profile for MODIS and Sentinel-1.

Cheers!

David
