### 2026-06-25 Digdgeo/Ndvi2Gif#5 @davemlz
JOSS Review: Available satellites for `NdviSeasonality`
Hi @Digdgeo!

This issue is related to your current submission to JOSS (https://github.com/openjournals/joss-reviews/issues/10654).

When trying to use another satellite, as described in the docs (https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#understanding-the-parameters), the code fails and shows a different set of options:

Cheers!

David
