### 2026-06-25 Digdgeo/Ndvi2Gif#6 @davemlz
JOSS Review: No `key` containment for `NdviSeasonality`
Hi @Digdgeo!

This issue is related to your current submission to JOSS (https://github.com/openjournals/joss-reviews/issues/10654).

When passing a random key to `NdviSeasonality` there is no containment:

Cheers!

David
