### 2026-06-25 2369342
Fix docs satellite codes, add key validation, prune empty doc pages
- NdviSeasonality: raise ValueError on invalid 'key' instead of silently
  defaulting to 'max' (JOSS review #6)
- Docs: replace non-existent 'L8'/'L9' satellite codes with 'Landsat' and
  list all supported sensors (S2, S1, Landsat, MODIS, S3, ERA5, CHIRPS) (JOSS review #5)
- Book: remove empty placeholder tutorials/use-cases/notebook from the ToC,
  delete orphan stub files and fix dangling cross-links (JOSS review #3)
