+++ book/tutorials/statistical_methods.md
@@ -1,5 +1,0 @@
-# Statistical_methods
-
-*This tutorial is under development and will be available soon.*
-
-Check back later or contribute via [GitHub](https://github.com/Digdgeo/Ndvi2Gif).
