+++ book/advanced/drought.md
@@ -1,5 +1,0 @@
-# Drought
-
-*This section is under development and will be available soon.*
-
-Check back later or contribute via [GitHub](https://github.com/Digdgeo/Ndvi2Gif).
