+++ book/notebooks/04_sar_vegetation.md
@@ -1,3 +1,0 @@
-# 04 sar vegetation
-
-*Notebook coming soon. In the meantime, check the [examples_notebooks](https://github.com/Digdgeo/Ndvi2Gif/tree/master/examples_notebooks) in the repository.*
