### 2026-06-25 Digdgeo/Ndvi2Gif#6 @Digdgeo
Hi @davemlz,

You were right. While "sat" and "index" were validated, the key (statistic) parameter silently fell back to "max" when given an invalid value. It now raises a ValueError listing the available options (max, min, median, percentile, mean, sum), consistent with the other parameters. Fixed in the repo; it will reach PyPI with the next release.

Thanks!

P.S. BTW, adding more statistic options is one of the easiest things on the roadmap — it'll come soon.
