### 2026-06-25 Digdgeo/Ndvi2Gif#5 @Digdgeo
Hi @davemlz,

Thanks for reporting this. This was a bug in the example code snippets: the documentation listed 'L8' and 'L9' as satellite codes, but those don't exist in the API — Landsat is accessed through a single 'Landsat' collection that merges Landsat 4, 5, 7, 8 and 9 (TM, ETM+, OLI and OLI-2).

I've corrected the satellite codes across the documentation, and the parameter list now reflects the actual supported sensors: S2, S1, Landsat, MODIS, S3, ERA5, CHIRPS. The docs site should be alredy refreshed.

Cheers!
