### 2026-06-25 Digdgeo/Ndvi2Gif#3 @Digdgeo
Hi @davemlz,

Thank you for the detailed list! I've started addressing it. Done so far:

- Removed the empty placeholder pages (the "under development" tutorials, use cases, and the SAR notebook) from the navigation and fixed the broken cross-links, so there are no more dead/empty pages in the docs. Honestly, I tend to be a bit too optimistic about my available time, and I'm adding new tutorials much more slowly than I'd like — so for now I'd rather show only the content that is actually ready and bring the rest back as I complete it.
- Fixed the L8/L9 satellite codes (see #5).

Still in progress (goal for this weekend): correcting the broken code examples in the Quick Start / ROI sections (the year/fps keyword mismatches, the multi-band palette and GeoJSON errors), completing the API reference so it matches the classes and methods in the repo (including SpatialPhenologyAnalyzer and SpatialTrendAnalyzer), publishing the Changelog and Contributing pages,  adding the index references, and linking the docs URL in the repo's About section.

I'll keep this issue updated as I tick items off.

Cheers!
