### 2026-06-25 8903a83
Print sensor status only for the selected satellite
Wrap the MODIS and Sentinel-1 status messages in their own self.sat
guards so the collection setup no longer prints processing profiles for
sensors other than the one in use (JOSS review #4).
