### 2026-06-25 Digdgeo/Ndvi2Gif#4 @Digdgeo
Hi @davemlz,

Thanks! Those prints were a bit of a leftover from development: as I was adding the different collections I left them in to double-check that each sensor was being configured correctly, and since they never really bothered me I ended up keeping them. But you're right that it makes much more sense only showing information relevant to the selected satellite.

I've cleaned this up — the status messages for each sensor (MODIS, Sentinel-1, etc.) are now only printed when that satellite is the one in use. It's already fixed in the repo and will be in PyPI with the next release.

Cheers!
