### 2026-06-25 Digdgeo/Ndvi2Gif#2 @Digdgeo
Hi @davemlz,

This reminds me of one of those "Yes, but..." jokes  ;P

I agree with you, actually after the redesign of the Analysis Capabilities box with your first comment I was trying to figure out which will be the best way to connect the outputs with the classes... And yes, that row can confuse as a tangled network rather than a set of independent outputs.

I'll fix it this weekend by grouping all the outputs (Maps, GIFs, Charts, Classification, Statistics) inside a single output box, with no arrows between them — just one arrow going to Download. I'll update the figure and regenerate the PDF.

Thanks again!
