### 2026-06-26 Digdgeo/Ndvi2Gif#6 @davemlz
Perfect! Closing this one!

Thanks, @Digdgeo!
