### 2026-06-26 Digdgeo/Ndvi2Gif#4 @davemlz
Perfect! Closing this one!

Thanks, @Digdgeo!
