### 2026-07-02 4a8a51a
Group outputs into a single box in the workflow figure (JOSS review #2)
