### 2026-07-02 Digdgeo/Ndvi2Gif#2 @Digdgeo
Hi @davemlz,

Sorry for the delay, I couldn't find time this weekend.

I've grouped all the outputs (Maps, GIFs, Charts, Classification, and Statistics) within a single Outputs box, with no arrows between them — just one arrow going down to Download. The tangled-network look should be gone now, and combined with the dashed border on the Analysis Capabilities box from your first comment,  I think the flow reads much more clearly.

The updated figure is already pushed to the repository (commit 4a8a51a) and I asked the bot to regenerate the paper PDF so you can re-check it.

Thanks again for your detailed review. I'm starting now with the last task (changes in docs)

Cheers :)
