### 2026-07-02 Digdgeo/Ndvi2Gif#2 @davemlz
Hi @Digdgeo!

It looks great! I will close this Issue now ;)

Cheers!

David
