### 2026-07-02 bfa810e
Fix and complete Jupyter Book docs (JOSS review #3)
- Correct all broken code examples (getting_started, tutorials, faq):
  get_year_composite/get_export/get_gif signatures, ROI code formats
  (s2:/wrs:/deimsid), band names, and cloud-mask parameters
- Rewrite the API reference with autodoc so it documents all classes and
  methods and stays in sync with the code
- Publish the Changelog and Contributing pages from the root files
- Add a title to the time series notebook; config cleanups (bib path,
  favicon, excluded stray files)
- Add index reference DOIs sourced from awesome-spectral-indices
[…]
