+++ book/reference/changelog.md
@@ -1,8 +1,2 @@
-# Changelog
-
-*This reference section is under development.*
-
-For now, see:
-- [GitHub Repository](https://github.com/Digdgeo/Ndvi2Gif)
-- [README](https://github.com/Digdgeo/Ndvi2Gif/blob/master/README.md)
-- [Contributing Guide](https://github.com/Digdgeo/Ndvi2Gif/blob/master/CONTRIBUTING.md)
+```{include} ../../CHANGELOG.md
+```
