### 2026-07-02 Digdgeo/Ndvi2Gif#3 @Digdgeo
Hi @davemlz,

Here's a summary of what's been addressed (everything is already pushed and the docs site has rebuilt):

Broken code examples:

- Fixed all the failing snippets in the Getting Started section: the get_year_composite() / get_export() / get_gif() calls no longer use non-existent keyword arguments (year, fps, etc.), the multi-band palette visualization is corrected, and the authentication test now filters a collection instead of a hard-coded (missing) image asset.
- Corrected the ROI code formats in roi_options (s2:<tile>, wrs:<path,row>, deimsid/<id>), which were causing the "Invalid GeoJSON geometry" errors.
- I also went ahead and fixed the same class of issues in the Tutorials and FAQ pages, so the whole book is consistent now.

API reference:

- Rewrote the API Reference using autodoc, so it now documents all classes and their methods directly from the source code and stays in sync automatically. This includes the previously missing SpatialTrendAnalyzer, SpatialPhenologyAnalyzer and HydroperiodAnalyzer.
- Added an overview table at the top and per-class sections, so each class is now navigable from the sidebar.
- The docs are now built from the repository source, so the API reference always matches the current code rather than the last PyPI release.

Missing pages:

- The Changelog and Contributing pages are now published (rendered from the repo files).
- Added the reference DOIs to the indices page, sourced from your Awesome Spectral Indices catalogue — thank you for that resource, I've added an acknowledgment note crediting it.
- Added the documentation URL to the repository's About section.

Other:

- I've removed the unfinished multi-sensor tutorial from the navigation for now; I'd rather bring it back (or a replacement) once it's actually complete.

I think that covers everything from your list — please let me know if I missed anything or if something still doesn't work during your review.

Thanks again!
