### 2026-07-04 Digdgeo/Ndvi2Gif#3 @davemlz
Hi @Digdgeo!

Great! Thank you for addressing all of the issues! I have encountered a bit more and added them to the list ;) Please let me know when you address them to continue with the review.

Cheers!

David
