### 2026-07-06 Digdgeo/Ndvi2Gif#7 @giswqs
JOSS Review: Invalid Notebook
https://github.com/Digdgeo/Ndvi2Gif/blob/master/docs/notebooks/Basic_Usage.ipynb

https://github.com/openjournals/joss-reviews/issues/10654
