### 2026-07-06 Digdgeo/Ndvi2Gif#8 @giswqs
JOSS Reivew: ee.Initialize() requires a project id
Most examples use `ee.Initialize()`, which does not work. A project ID is required to initialize EE.

https://github.com/openjournals/joss-reviews/issues/10654
