### 2026-07-06 Digdgeo/Ndvi2Gif#9 @giswqs
JOSS Review: GitHub Actions outdated
The GitHub Actions are using deprecated packages, including but not limited to:

- Python 3.10
- actions/checkout@v3
- actions/checkout@v4
- actions/setup-python@v4

https://github.com/openjournals/joss-reviews/issues/10654
