### 2026-07-06 Digdgeo/Ndvi2Gif#10 @giswqs
JOSS Review: Community guidelines
The Community guidelines can be improved. It does not include forking the repo, creating new branch, etc.

https://digdgeo.github.io/Ndvi2Gif/reference/contributing.html#development-setup

https://github.com/openjournals/joss-reviews/issues/10654
