### 2026-07-06 Digdgeo/Ndvi2Gif#11 @giswqs
JOSS Review: Paper title
`via Google Earth Engine` is not commonly used in paper titles. Suggest changing to `Using Google Earth Engine` or `with Google Earth Engine`.

https://github.com/openjournals/joss-reviews/issues/10654
