### 2026-07-06 Digdgeo/Ndvi2Gif#12 @giswqs
JOSS Review: Software paper summary section
The second paragraph of the Summary section is quite technical, not well suited for non-specialist audience.

- [ ] Summary (required section): Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

https://github.com/openjournals/joss-reviews/issues/10654
