### 2026-07-06 Digdgeo/Ndvi2Gif#12 @giswqs
Also, the phrase `machine learning land cover classification` sounds awkward. It should probably be `machine learning-based land cover classification`. Since almost all land cover classification methods are based on machine learning algorithms, it can probably be further simplified as `land cover classification`.
