### 2026-07-06 b14105f
Address JOSS review round 2 (reviewers David and Qiusheng)
Docs / reproducibility (David):
- quick_start: shrink export ROI and document the ~50 MB synchronous
  GeoTIFF download limit, pointing large exports to Drive
- time_series / classification: define the roi used in the quick-start
  snippets so they run standalone
- water_quality: load the reservoir ROI from a public Earth Engine asset
  (projects/ee-digdgeografo/assets/CHG_CaptacionesEmbalses) instead of a
  missing local shapefile; same for the companion notebook
- translate notebook 02 (land cover classification) to English

Earth En
[…]
