+++ docs/tutorials/ndvi2gif.md
@@ -26,7 +26,7 @@ import ee
 import geemap
 from ndvi2gif import NdviSeasonality
 
-ee.Initialize()
+ee.Initialize(project='your-project-id')
 
 roi = ee.Geometry.Rectangle([-6.55, 36.85, -6.10, 37.20])
 
