### 2026-07-06 Digdgeo/Ndvi2Gif#3 @Digdgeo
Hi @davemlz, thanks a lot one more time!

All five are fixed in commit Digdgeo/Ndvi2Gif@b14105f, and the documentation site has been rebuilt and redeployed.

**1. GeoTIFF export exceeds the 50 MB limit**
`get_export()` downloads directly from Earth Engine, whose synchronous requests are capped at ~50 MB, so the Madrid example (12 monthly bands over a large ROI) was too big. I shrank the quick-start ROI so the example runs end-to-end, and added a note in [Export as GeoTIFF](https://digdgeo.github.io/Ndvi2Gif/getting_started/quick_start.html#export-as-geotiff) explaining the limit and pointing large exports to `export_to_drive()` (anasynchronous task with a much larger budget).

**2 & 3. `NameError: name 'roi' is not defined`**
The quick-start snippets in [Time Series] (https://digdgeo.github.io/Ndvi2Gif/advanced/time_series.html#quick-start) and Classification] (https://digdgeo.github.io/Ndvi2Gif/advanced/classification.html#quick-start) now define `roi` (a self-contained ee.Geometry.Rectangle`) before it is used, so they run standalone.

**4. `DataSourceError`: reservoir shapefile not found**
I've uploaded the CHG reservoirs layer a asset** (`projects/ee-digdgeografo/assets/CHG_CaptacionesEmbalses`), and the [Water Quality](https://digdgeo.github.io/Ndvi2Gif/advp-and-roi) page (and its companion notebook) now loads it directly from that asset — no local file needed, fully reproducible.

**5. Notebook 02 partially in Spanish**
[`02_landcover_classification`](https://digdgeo.github.io/Ndvi2Gif/notebooks/02_landcover_classification.html) is now fully in English.

I wish I could be sure there will be no more bugs :/

Cheers!
