### 2026-07-06 Digdgeo/Ndvi2Gif#7 @Digdgeo
Fixed in Digdgeo/Ndvi2Gif@b14105f.

This was an issue after re-saving the notebook in Google Colab. It now validates with `nbformat` and renders correctly on GitHub.

Thanks!
