### 2026-07-06 Digdgeo/Ndvi2Gif#8 @Digdgeo
Fixed in Digdgeo/Ndvi2Gif@b14105f.

All examples now use `ee.Initialize(project='your-project-id')` across the docs and the README, and I added a note in the [Authentication](https://digdgeo.github.io/Ndvi2Gif/getting_started/authentication.html) page about the Cloud project ID being required.

Thanks!
