### 2026-07-06 651c752
Bump GitHub Pages actions to v5 (JOSS review #9)
upload-pages-artifact@v3 -> @v5 and deploy-pages@v4 -> @v5, so the
deploy workflow no longer uses any outdated action versions.
