### 2026-07-06 Digdgeo/Ndvi2Gif#9 @Digdgeo
Fixed in Digdgeo/Ndvi2Gif@b14105f and Digdgeo/Ndvi2Gif@651c752.

Updated all workflows to current, non-deprecated versions:

- `actions/checkout@v3` / `@v4` → `actions/checkout@v5`
- `actions/setup-python@v3` / `@v4` → `actions/setup-python@v6`
- `actions/upload-pages-artifact@v3` → `@v5`
- `actions/deploy-pages@v4` → `@v5`
- replaced the pinned **Python 3.10** with a `3.11` / `3.12` / `3.13` test matrix.

The documentation-deploy workflow ran successfully with all the updated versions.

Thanks a lot for pointing this out!
