### 2026-07-06 Digdgeo/Ndvi2Gif#10 @Digdgeo
Fixed in Digdgeo/Ndvi2Gif@b14105f.

The [contributing guidelines](https://digdgeo.github.io/Ndvi2Gif/reference/contributing.html) now include a full **fork → branch → pull request** workflow (with the `upstream` remote and a PR checklist).

Thanks again :)
