### 2026-07-06 Digdgeo/Ndvi2Gif#11 @Digdgeo
Fixed in Digdgeo/Ndvi2Gif@b14105f.

Changed the title to *"Ndvi2Gif: A Python Package for Multi-Seasonal Remote Sensing Analysis **with** Google Earth Engine"*. The regenerated PDF already reflects this.

Thanks for the suggestion!
