### 2026-07-06 Digdgeo/Ndvi2Gif#3 @davemlz
Hey @Digdgeo!

Great work there! I know sometimes it's a pain, but you are doing amazing! Thanks for addressing all of the issues. Closing this one ;)

Cheers!

David
