### 2026-07-06 Digdgeo/Ndvi2Gif#9 @giswqs
The latest version is `actions/checkout@v7`. `actions/checkout@v5` may be deprecated soon.
