### 2026-07-06 Digdgeo/Ndvi2Gif#12 @Digdgeo
Fixed in Digdgeo/Ndvi2Gif@b14105f.

I've rewroted the second paragraph of the Summary thinking in a non-specialist audience, removing the API/code jargon (`get_year_composite()`, `ee.ImageCollection`, `periods=12`, `collection.select(...).max()`, `get_stats()`) and describing the core idea in plain language with concrete examples.

I also applied your phrasing suggestion in the first paragraph: *"machine learning land cover classification"* → *"land cover classification"*. The regenerated PDF reflects both changes.

Thanks!
