### 2026-07-06 Digdgeo/Ndvi2Gif#13 @davemlz
JOSS Review: Publish a new version
Hi @Digdgeo!

This issue is related to your current submission to JOSS (https://github.com/openjournals/joss-reviews/issues/10654).

I think the last thing to do for this review (on my side) will be publishing a new version to update the printing already discussed in #4 ;)

Cheers!

David
