### 2026-07-06 f3be4b3
Remove legacy Sphinx docs and ReadTheDocs config
The Jupyter Book (book/, published on GitHub Pages) is now the single
source of documentation. The docs/ folder held an older Sphinx site that
duplicated it and was the source of several review issues (e.g. the
invalid notebook in docs/notebooks/).

- keep the newer, more complete Basic_Usage notebook in examples_notebooks/
  (the docs/ Hydroperiod and LandCover_Class notebooks were duplicates of
  ones already there)
- drop .readthedocs.yaml (the RTD build was orphaned; nothing links to it)
