+++ docs/api/ndvi2gif.md
@@ -1,7 +1,0 @@
-# API: ndvi2gif.ndvi2gif
-
-```{automodule} ndvi2gif.ndvi2gif
-:members:
-:undoc-members:
-:show-inheritance:
-```
