+++ docs/api/clasification.md
@@ -1,7 +1,0 @@
-# API: ndvi2gif.clasification
-
-```{automodule} ndvi2gif.clasification
-:members:
-:undoc-members:
-:show-inheritance:
-```
