+++ docs/api/timeseries.md
@@ -1,7 +1,0 @@
-# API: ndvi2gif.timeseries
-
-```{automodule} ndvi2gif.timeseries
-:members:
-:undoc-members:
-:show-inheritance:
-```
