+++ docs/requirements-doc.txt
@@ -1,4 +1,0 @@
-sphinx>=7
-myst-nb>=1.0
-sphinx-copybutton
-furo

[…]
