### 2026-07-06 Digdgeo/Ndvi2Gif#13 @Digdgeo
Thanks :)

I think I'll hold the new release until @giswqs has had the chance to check whether anything else needs changing, so that a single new version bundles all the review fixes at once. As soon as I have his final OK, I'll cut the release and publish it to PyPI. BTW, I also should upload it to Zenodo, shouldn't I?

Thanks again for your help during the whole process!

Cheers,
Diego
