### 2026-07-06 Digdgeo/Ndvi2Gif#13 @davemlz
Linking it to Zenodo would be great, you have to do the linking before a publish so the published version gets the Zenodo repo and the DOI ;)

Happy to review this amazing package!
