### 2026-07-06 Digdgeo/Ndvi2Gif#13 @giswqs
I have completed my review. Looks good to me.
