### 2026-07-06 e361c9f
Add community engagement paragraph and AET citation (JOSS review, Collaborative effort)
- Research Impact Statement: new paragraph on community uptake (downloads,
  forks/stars, dissemination, user-driven development, studies in preparation)
- Cite AET 2026 congress presentation (@GarciaDiaz2026) in paper.bib
- Update affiliation to LAST-EBD, CSIC
