+++ paper/paper.bib
@@ -81,6 +81,18 @@
   doi={10.4995/raet.2023.18767}
 }
 
+@inproceedings{GarciaDiaz2026,
+  title={Ndvi2Gif: Análisis multi-estacional mediante Google Earth Engine},
+  author={García-Díaz, Diego},
+  booktitle={XXI Congreso de la Asociación Española de Teledetección (AET): Teledetección Multidisciplinar: Miradas Diversas, Soluciones Comunes},
+  series={Congresos UEx},
+  volume={2},
+  number={2},
+  year={2026},
+  publisher={Universidad de Extremadura, Servicio de Publicaciones},
+  doi={10.17398/3101-7177.2.100}
+}
+
 @techreport{DiazDelgado2024b,
   title={eLTER PLUS Deliverable D4.4 - Piloting of eLTER specific downstream services contributing to eLTER Standard Observation Variables},
   author={Díaz-Delgado, Ricardo and García-Díaz, Diego and Marangi, Catia and Silver, Michael and Oggioni, Alessandro and Tagliolato Acquaviva D'Aragona, Paolo and Karnieli, Arnon and Peterseil, Johannes and Vicario, Saverio and Anttila, Sanna and Böttcher, Kristin and Bolton, William and Wohner, Christoph and Minic, Vojislava},
