### 2026-07-07 facb98a
Release v1.3.1: quiet per-sensor status prints (JOSS #4); add Zenodo and CITATION metadata
