+++ CHANGELOG.md
@@ -9,6 +9,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2026-07-07
+
+### Fixed
+
+- Sensor status messages during collection setup are now printed only for the **selected** satellite, instead of for every supported sensor, reducing console verbosity (JOSS review, davemlz #4).
+
+### No Breaking Changes
+
+Full backward compatibility with v1.3.0.
+
 ## [1.3.0] - 2026-06-14
 
 ### Added
