+++ CITATION.cff
@@ -1,0 +1,14 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it using these metadata."
+title: "Ndvi2Gif: A Python Package for Multi-Seasonal Remote Sensing Analysis with Google Earth Engine"
+type: software
+authors:
+  - family-names: "García Díaz"
+    given-names: "Diego"
+    orcid: "https://orcid.org/0000-0002-2757-7112"
+    affiliation: "Laboratorio de SIG y Teledetección de la Estación Biológica de Doñana (LAST-EBD), CSIC, Spain"
+repository-code: "https://github.com/Digdgeo/Ndvi2Gif"
+url: "https://digdgeo.github.io/Ndvi2Gif/"
+license: MIT
+version: 1.3.1
+date-released: "2026-07-07"
