### 2026-07-07 Digdgeo/Ndvi2Gif#13 @Digdgeo
Hi @davemlz and @giswqs,

I've just published v1.3.1, which ships the fix for the print output discussed in #4 (sensor status is now printed only for the selected satellite). It's already live on PyPI and it will also be available on conda-forge shortly:

https://pypi.org/project/ndvi2gif/1.3.1/

The repo was already linked to Zenodo, so this release got archived automatically. The v1.3.1 archive DOI is: 10.5281/zenodo.21246724 (concept DOI: 10.5281/zenodo.20693553)

Since @giswqs had already completed his review, this single release bundles all the review fixes at once, as I mentioned earlier.

I really appreciate all your help and feedback during the process — it really made Ndvi2Gif better!

Cheers,
Diego
