### 2026-07-08 Digdgeo/Ndvi2Gif#13 @davemlz
Hey @Digdgeo! Great! Closing this one now then! 😉
