### 2026-07-09 Digdgeo/Ndvi2Gif#14 @mengqi-z
JOSS Paper
Hello @Digdgeo,

This is the proofreading of the JOSS paper: openjournals/joss-reviews#10654

Could you please address the following points before we can proceed with the review process? The line numbers below refer to the PDF version of the manuscript.

1. Line 12: Please define SAR by spelling out its full name the first time it appears, if it is an abbreviation for a technical term.
2. Please format all software, model, package, and function names using code style in `paper.md` (e.g., \`Ndvi2Gif\`, \`geemap\`, \`eemont\`). Ensure this formatting is applied consistently throughout the paper wherever applicable.
3. Line 64: Please define ROI.

Thank you.
