### 2026-07-10 f36986d
Address JOSS editor proofreading (#14): define SAR and ROI, apply code style to software names
