### 2026-07-10 Digdgeo/Ndvi2Gif#14 @Digdgeo
Hi @mengqi-z,

Thank you for the proofreading. I've addressed all three points in `paper.md` (already pushed to master):

- **Line 12 — SAR:** now spelled out as "Synthetic Aperture Radar (SAR)" on its first appearance in the text.
- **Code style:** all software, package and function names (`Ndvi2Gif`, `geemap`, `eemont`, `wxee`, `geeltermap`, etc.) are now formatted in code style consistently throughout the paper.
- **Line 64 — ROI:** now defined as "region of interest (ROI)" on its first use.

Separately, following @davemlz's suggestion during the review, I published v1.3.1, which ships the per-sensor status-print fix he requested. It is now live on PyPI and conda-forge and archived on Zenodo:

- Archive DOI: 10.5281/zenodo.21246724
- Version: v1.3.1

Please let me know if anything else is needed.

Thanks a lot for your help during the process :)
