### 2026-08-18 #10654 @kthyng
@Digdgeo Sorry for my delay. You can go ahead with any subsequent changes to your repo.

Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "earth" should be capitalized but please check for others.
