### 2026-08-19 4f6d0ca
Protect capitalization in paper.bib references (JOSS AEiC request)
Wrap proper nouns and lowercase software names in braces so citeproc
does not sentence-case them: Google Earth Engine, Sentinel-1, DEIMS-SDR,
Python, Earth Observation, Doñana, eLTER, geemap, eemont, wxee, xarray,
PhenoApp and Ndvi2Gif.
