+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Gorelick2017,
-  title={Google Earth Engine: Planetary-scale geospatial analysis for everyone},
+  title={{Google Earth Engine}: {P}lanetary-scale geospatial analysis for everyone},
   author={Gorelick, Noel and Hancher, Matt and Dixon, Mike and Ilyushchenko, Simon and Thau, David and Moore, Rebecca},
   journal={Remote Sensing of Environment},
   volume={202},
@@ -10,7 +10,7 @@
 }
 
 @article{Wu2020,
-  title={geemap: A Python package for interactive mapping with Google Earth Engine},
+  title={{geemap}: {A} {Python} package for interactive mapping with {Google Earth Engine}},
   author={Wu, Qiusheng},
   journal={Journal of Open Source Software},
   volume={5},
@@ -22,7 +22,7 @@
 }
 
 @article{Wohner2019,
-  title={DEIMS-SDR--A web portal to document research sites and their associated data},
+  title={{DEIMS-SDR}--{A} web portal to document research sites and their associated data},
   author={Wohner, Christoph and Peterseil, Johannes and Oggioni, Alessandro and Bertrand, Nicolas and Schentz, Herbert},
   journal={Ecological Informatics},
   volume={51},
@@ -33,7 +33,7 @@
 }
 
 @article{Vollrath2020,
-  title={Angular-based radiometric slope correction for Sentinel-1 on Google Earth Engine},
+  title={Angular-based radiometric slope correction for {Sentinel-1} on {Google Earth Engine}},
   author={Vollrath, Andreas and Mullissa, Adugna and Reiche, Johannes},
   journal={Remote Sensing},
   volume={12},
@@ -45,7 +45,7 @@
 }
 
 @article{Montero2021,
-  title={eemont: A Python package that extends Google Earth Engine},
+  title={{eemont}: {A} {Python} package that extends {Google Earth Engine}},
   author={Montero, David},
   journal={Journal of Open Source Software},
   volume={6},
@@ -58,13 +58,13 @@
 
 @software{Zuspan2021,
   author={Zuspan, Aaron},
-  title={wxee: A Python interface between Earth Engine and xarray},
+  title={{wxee}: {A} {Python} interface between {Earth Engine} and {xarray}},
   url={https://github.com/aazuspan/wxee},
   year={2021}
 }
 
 @techreport{DiazDelgado2024,
-  title={In-situ validation of Land Surface Phenology, Land Surface Temperature and Surface Water derived from Earth Observation products: Doñana protected area as a potential cal/val supersite},
+  title={In-situ validation of {Land Surface Phenology}, {Land Surface Temperature} and {Surface Water} derived from {Earth Observation} products: {Doñana} protected area as a potential cal/val supersite},
   author={Díaz-Delgado, Ricardo and García-Díaz, Diego},
   institution={eLTER Plus Project},
   year={2024},
@@ -72,7 +72,7 @@
 }
 
 @article{Garcia2023,
-  title={PhenoApp: A Google Earth Engine based tool for monitoring phenology},
+  title={{PhenoApp}: {A} {Google Earth Engine} based tool for monitoring phenology},
   author={García-Díaz, Diego and Díaz-Delgado, Ricardo},
   journal={Revista de Teledetección},
   number={61},
@@ -82,9 +82,9 @@
 }
 
 @inproceedings{GarciaDiaz2026,
-  title={Ndvi2Gif: Análisis multi-estacional mediante Google Earth Engine},
+  title={{Ndvi2Gif}: {A}nálisis multi-estacional mediante {Google Earth Engine}},
   author={García-Díaz, Diego},
[…]
