### 2026-08-19 2f2e6a7
Unify eLTER PLUS project name in paper.bib institution fields
