+++ paper/paper.bib
@@ -66,7 +66,7 @@
 @techreport{DiazDelgado2024,
   title={In-situ validation of {Land Surface Phenology}, {Land Surface Temperature} and {Surface Water} derived from {Earth Observation} products: {Doñana} protected area as a potential cal/val supersite},
   author={Díaz-Delgado, Ricardo and García-Díaz, Diego},
-  institution={eLTER Plus Project},
+  institution={eLTER PLUS Project},
   year={2024},
   doi={10.62880/rars25009}
 }
