### 2026-08-19 #10654 @Digdgeo
Hi @kthyng,

Great news, thanks a lot for picking this up!

I've fixed the capitalization in the references: proper nouns and software names are now protected with braces in paper.bib (Google Earth Engine, Earth Observation, Sentinel-1, DEIMS-SDR, Python, Doñana, eLTER, Land Surface Phenology/Temperature, and the lowercase package names geemap, eemont, wxee and xarray, which were being capitalized at the start of a title). I've regenerated the PDF and the reference list now renders correctly.

About subsequent changes to the repo: thanks for the green light. I do have some new features ready, but since the paper seems to be close to publication I'd rather wait and cut the new release once it's out, so that v1.3.1 and the Zenodo archive (10.5281/zenodo.21246724) reported here remain the ones matching the publication. The commits I've just pushed only touch paper.bib and the compiled paper.pdf — no changes to the package code — so the version and archive stay valid.

Thanks again!
